Fixed some cms issues
