--- content/portfolio_al/arsimtaret-ne-maqedoni-kane-rroga-me-te-uleta.md
@@ -1,54 +1,0 @@
----
-title: Arsimtarët në Maqedoni kanë rroga më të ulëta dhe shpenzime më të larta për jetesë
-cover: '/images/uploads/salary.jpg'
-date: 2020-01-26
-author: ['Rrita Syla', 'Ardit Ramadani']
-published:
- Diversity Media: 'http://diversitymedia.mk/%d0%bd%d0%b0%d1%81%d1%82%d0%b0%d0%b2%d0%bd%d0%b8%d1%86%d0%b8%d1%82%d0%b5-%d0%b2%d0%be-%d0%bc%d0%b0%d0%ba%d0%b5%d0%b4%d0%be%d0%bd%d0%b8%d1%98%d0%b0-%d0%b8%d0%bc%d0%b0%d0%b0%d1%82-%d0%bd%d0%b0%d1%98/'
-
----
-
-Shkruajnë: Rrita Syla & Ardit Ramadani
-
-Në xhirollogarinë e dy prindërve të punësuar në arsimin fillor në RMV, në fund të muajit hyjnë gjithsej 800 Euro, në Shqipëri 1,000 Euro kurse në Kosovë 1,098 Euro. Mësimdhënësit ankohen për atë se nuk mund ta kalojnë muajin me këto paga kurse sipas ekspertëve, në Maqedoninë e Veriut përpos që pagat janë më të ulëta, shpenzimet janë më të larta në krahasim me vendet e lartpërmendura.
-
-*Maqedonia*
-
-Paga mujore për një mësimdhënës në arsimin fillor në Maqedoninë është 400 Euro që rezulton se një familje e cila ka të punësuar dy mësimdhënës në arsimin fillor, të hyrat mujore i ka 800 Euro, kurse për shportën mujore të harxhimeve për një familje me katër anëtarë, sipas Lidhjes të Sindikatave të Maqedonisë së Veriut, nevojiten 534 Euro.
-Duke minusuar harxhimet mujore, një familje e cila ka të punësuar dy mësimdhënës në arsimin fillor arrin që në fund të muajit të ketë 266 Euro të mbetura për mjete shtesë jo të domosdoshme për ushqim dhe fatura.
-Në qoftë se grumbullojmë mjetet e fituara nga njëri prind i punësuar në arsim me prindin tjetër i cili punn një punë me pagë minimale, si mjete shtesë kësaj familje i ngelin 106 Euro.
-
-*Shqipëri*
-
-Paga mujore për mësimdhënësit e arsimit fillor në Republikën e Shqipërisë është (500 Euro), andaj sipas kësaj në xhirollogarinë e familjes e cila ka të punësuar dy mësimdhënës, në fund të muajit arrinë (1000 Euro) kurse për të kaluar muajin një familje me katër anëtarë, sipas Entit statistikor të Shqipërisë Instat, nevojiten 619 Euro.
-Në qoftë se mënjanojmë harxhimet mujore nga totali i të ardhurave mujore për këtë familje si mjete për shpenzime shtesë ngelin 381 Euro.
-
-Nga ana tjetër, në qoftë se kemi të bëjmë me një familje katëranëtarëshe, në të cilën njëri prind është i punësuar në arsim kurse tjetri bën një punë e cila paguhet minimalisht,
-
-![Salary](/images/uploads/salary-al.png)
-
-*Kosova*
-
-Një mësimdhënës në Republikën e Kosovës për çdo muaj merr 549 Euro, në llogarinë e një familjeje e cila ka të punësuar dy mësimdhënës në arsimin fillor, arrijnë 1,098 Euro. Nga ana tjetër, në mënyrë që një familje me katër anëtarë ta kaloje muajin pa problem, sipas Agjencisë për statistikë të Kosovës nevojiten 650 Euro.
-Duke larguar harxhimet e nevojshme mujore për një familje me katër anëtarë në Kosovë, si mjete për shpenzime shtesë ngelin 448 Euro.
-
-E në qoftë se njëri nga prindët bën një punë me pagë minimale kurse tjetri është i punësuar në arsim, si mjete shtesë apo jo të domosdoshme ju ngelin 69 Euro.
-
-![Salary](/images/uploads/salary-al-2.png)
-
-Sipas mësimdhënëses Gonxhe Mehmedi paga të cilën ajo e merr nuk i mjafton për të jetuar. Sa i takon rritjes së pagave në arsim, ajo vlerëson se kjo rritje e vogël nuk është e mjaftueshme.
-“Të jesh mësimdhënës në Maqedoni nuk është aspak e lehtë, duke filluar që nga problemet teknike deri tek paga më e ulët në rajon, e cila nuk mjafton as për të mbuluar pagesat e shpenzimeve mujore dhe që rrezikon të ardhmen e këtij profesioni. Edhe pse, gjatë viteve të kaluara ka pasur rritje të vogla të rrogave, përsëri nuk kënaqin nevojat elementare në krahasim me çmimet e larta” theksoi Gonxhe Mehmedi.
-
-![Salary](/images/uploads/salary-al-3.png)
-
-Për eksperten e statistikës Shpresa Syla, të dhënat e siguruara nga ne janë shqetësuese për arsye se sipas saj këto shifra vetëm se ndikojnë në demotivimin e mësidhmënësve, ajo bën thirrje deri tek Sindikata e Arsimit të këmbëngulin për rritjen e pagave.
-
-*“Në qoftë se analizohen pagat e familjeve që kanë dy mësimdhënës në arsim , ne vendet e rajonit , vërejmë se në RMV, përpos që janë pagat më të ulëta ,nga njëra anë , nga ana tjetër shpenzimet për jetesë janë më të larta. Sipas të dhënave të paraqitura për rrogat, në Maqedoni-66.75%; Shqipëri-59% dhe Kosovë 61.9% i dedikohen shportës së shpenzimeve.
-Inkurajojmë sindikatën e arsimit të këmbëngulë ne rritjen e pagave – një rritje e pagave prej 10% do të ndikonte ne rritje te mjeteve te mbetura për shpenzime nga 33.25% në 39,31%. Në të kundërtën arsimi i mesëm obligativ , pamundësinë në zhvillim të karrierës, pagat e ulëta do të ndikonin në demotivimin e arsimtarëve dhe rritjen në vazhdimësi të cilësisë në arsim”* tha Shpresa Syla.
-
-Marrë parasysh këto, mund të konstatojmë se një familje me një rrogë mësimdhënësi dhe një rrogë tjetër minimale as në Kosovë, as në Maqedoninë e Veriut e as në Shqipëri nuk mund t’ia dalë për shpenzime tjera shtesë pasi që, sipas konstatimeve tona, mjetet shtesë të cilat ju ngelin janë shumë të vogla, edhe atë: Kosovë 69 Euro, Shqipëri 92 Euro, Maqedonia e Veriut 106 Euro.
-Sipas kësaj, mund të thuhet se një familje në të cilën punon vetëm njëri prind, edhe atë në arsim nuk mund t’i mbulojë as shpenzimet e domosdoshme mujore për jetesë.
-
-*Me termin mjete shtesë apo shpenzime shtesë nënkuptojmë ato mjete të cilat ju ngelin pasi të paguhen shpenzimet themelore për jetesë. Këtu hyjnë pushimet apo nevojat tjera personale të cilat nuk përfshihen në shpenzimet mujore.
-
-Burime: Kosovo Agency of Statistics, Lidhja e sindikatave të Maqedonisë , Enti shtetëror i statistikës, Enti statistikor të Shqipërisë Instat, Eurostat.
