--- content/portfolio_al/jovana-georgieva-mbajtese-medaljeje-ne-karate-sfida-me-e-madhe-eshte-qe-frika-te-shnderrohet-ne-liri.md
@@ -1,33 +1,0 @@
----
-title: 'Jovana Georgieva, mbajtëse medaljeje në karate: Sfida më e madhe është që frika të shndërrohet në liri'
-cover: '/images/uploads/jovana.jpg'
-date: 2020-02-21
-author: ['Teodora Noveska', 'Shefkije Alasani']
-published:
- Diversity Media: 'http://diversitymedia.mk/jovana-georgieva-mbajtese-medaljeje-ne-karate-sfida-me-e-madhe-eshte-qe-frika-te-shnderrohet-ne-liri/'
----
-
-Shkruajnë: Teodora Noveska dhe Shefkije Alasani
-
-„Kur kisha pesë vjet, babai më çoi në karate, mbeta atje që atëherë. Nuk kisha fëmijëri tipike, vazhdimisht isha në stërvitje, besoja në ëndrra, kërkoja medalje, ballafaqohesha me frika që kur di për veten”
-
-E buzëqeshur, gazmore dhe komunukative, Jovana Georgieva është kampione seniore shtetërore dhe mbajtëse e medaljes bronze evropiane në karate. Të jesh sportist, sipas saj, është privilegj i madh, ndërsa karateja është dashuria e saj më e madhe.
-
-Kur ka qenë fëmijë, ka qenë çapkëne dhe hiperaktive, është interesuar për gjithçka, ka dashur të dijë shumë, ndërsa energji thotë se ka pasur po aq sa sot i mungon.
-
-“Babai më çoi në karate kur kisha 5 vjet. Për dore! Dhe ashtu mbeta atje. Viti 2020 është viti i moshës time madhore në karate, do të thotë ushtroj 18 vjet. Fillova në klubin e karatesë “Astibo”, pjesën më të madhe e kalova në “Makedonka”, ndërsa tani ushtroj në “Olimpik”. Nuk kisha fëmijëri tipike. Vazhdimisht isha në stërvitje. Besoja në ëndrra, kërkoja medalje. Ballafaqohesha me frika që kur di për veten. Dëshira për karate ishte dëshirë fëmijërore, por me kalimin e viteve kalonte në formë më të ndryshme – mbizotërimi mbi veten, përballja me veten, dëshira për të shtyrë veten të ec përpara. Karateja më është leksioni më i madh jetësor. Dashuria dhe dëshira të çojnë përpara, por rrugën e trason qëndrueshmëria dhe besimi në vete”, sqaron Jovana.
-
-Duke qenë ulur përballë meje, më jep përshtypjen e një gruaje të fuqishme, por me lëvizje gracioze. Lehtë mund të bëhet idhull i kujtdo. E pyes nëse ka ndonjë idhull të saj apo dikë si model. Thotë se nuk ka idhull. Ka gra të cilat i admiron, por e kupton se edhe vet po bëhet një e tillë dhe krenohet me këtë. Inspirimi i vetëm i është motivimi për të tejkaluar veten dhe frikën ta shndërrojë në liri. Përgjigjja akoma më shumë më përforcon imazhin për të dhe fillova ta admiroj.
-
-Është kampione e lartë shtetërore. Ka medalje bronze evropiane në vitin 2017 dhe bronze ballkanike në po të njëjtin vit.
-
-“Të jesh sportist është një privilegj i madh. Por, kjo arrihet si rezultat i sakrificës së vazhdueshme që e bëni për të qenë në formën tuaj maksimale fizike dhe psikike. Pengesa ka pafund. Pengesa ime e parë është se dita ka 24 orë. Çdo leksion tjetër i karatesë më ka qenë vetëm lehtësim. Ndjenja e të qenit mbajtës i medaljes evropiane është diçka që nuk mund të përshkruhet me fjalë. Paramendo, me vite i shikon ata persona që fitojnë. Të janë fiktivë. Dhe papritmas, pas 10 vitesh stërvitjeje, pa asnjë medalje dhe pa shpresë, gjatë vitit të parë të studimeve në fakultet, unë fitoj medaljen ballkanike, mandej evropiane dhe jam mes tyre. Ndjenja është e pabesueshme. Gjithçka për të cilën ke pasur mall që si fëmijë, të realizohet. Nuk je i vetëdijshëm se kjo është ajo për të cilën çdo ditë ke bërë gjithçka. Secila ngritje në tetë – merr kuptim. Çdo goditje dhe çdo mendim që ia ke kushtuar ëndrrës”, thotë Jovana.
-
-Më ishte interesant fakti se në Kampionatin Evropian e mundi kampionen e deriatëhershme evropiane. E pyeta se si ndihet lidhur me këtë. Përvoja, thotë, ka qenë e ndryshme, sepse nuk i ka besuar vetes aq shumë, sa që i është dorëzuar momentit dhe përqendrimin e ka pasur shumë të lartë. Ka dashur ta befasojë veten dhe ia ka dalë mbanë.
-Është viti i katërt, studion biologji molekulare në FMN. Fakultet i cili, sipas shumicës së njerëzve që i pyesni, është i vështirë dhe kërkon shumë vëmendje. Nuk e di nëse edhe asaj i duket ashtu, por flet me shumë lehtësi dhe entuziazëm të madh për përvojën në fakultet.
-“Fakultetin e kam dashuri. Si edhe çdo gjë tjetër që punoj. Për arsye se, gjithçka që bëj e bëj me dashuri. E kam shumë të vështirë të jetoj në ditë me 24 orë, sepse dua të arrij dhe të rregulloj gjithçka. Të mësoj, të ushtroj, të lexoj dhe të fle. Jeta ime sociale pëson shumë, por kohës së fundit e kuptoj se nuk po humb shumë, jo, përderisa gjithçka që bëj më shpërblen dhe më mbush me energji. E dua natyrën dhe faktin se si format e ndryshme të jetës janë të përsosura, si energjia ka gjetur rrjedhë me materien. Fakulteti më zbret në tokë. Nuk është lehtë të kombinohet me stërvitje, por në fund të fundit, kur e din se sa bën për vete, ajo të bën të lumtur”, sqaron Jovana.
-
-Mbarojmë me bisedën për karatenë. E pyes se si është situata në Maqedoni.
-“Maqedonia ka karate të shkëlqyer, shumë konkurruese me fuqitë e mëdha. Tanimë kemi emra botërorë. Të nisim nga Pavllovi medaljet e seniorëve, mandej medaljet e të rinjve, medaljen olimpike të Veselit, medaljet evropiane të Smillkovskit, Svinarskit. Botërori Bojan Kostov. Të rejat Mileva dhe Bozhinovska me medalje ballkanike. Vajza për respekt e cila lidh mjekësinë dhe karatenë, kampionia evropiane Veronika Kostovska. Maqedonia ka talente të jashtëzakonshme, që jam e lumtur që i kam pranë vetes”, e përfundon bisedën Jovana Georgieva.
-
-Përshëndetemi me Jovanën dhe në kokë më tingëllon motoja e saj: “Urat që do djeg, le të ma shndrisin rrugën”.
