--- content/portfolio_al/korrupsioni-dhe-ryshfeti-eshte-nje-shprehje.md
@@ -1,45 +1,0 @@
----
-title: Korrupsioni dhe ryshfeti është një shprehje e pafuqisë intelektuale
-cover: '/images/uploads/article-4.jpg'
-date: 2020-01-26
-author: ['Maja Terziova', 'Adem Jashari']
-published:
- Diversity Media: 'http://diversitymedia.mk/korrupsioni-dhe-ryshfeti-eshte-nje-shprehje-e-pafuqise-intelektuale/'
----
-
-Shkruajnë: Maja Terziova dhe Adem Jashari
-
-„Personi i cili situatat konfliktuoze i krijon vetë dhe janë nën regjinë e tij, i ka ushqim të patjetërsueshëm dhe shpirtëror, me mendje të robëruar dhe shumë shpesh vuajnë më shumë persona që ndodhen pranë tij. Asnjëherë nuk tregon dëshirë dhe prirje për mirëkuptim në marrëdhëniet ndërnjerëzore, respekt dhe vëmendje të ndërsjellë, nuk e ka asnjëherë ndjenjën e fajit, përkundrazi – në çdo situatë konflikti gjithnjë di ta fajësojë palën tjetër dhe të hakmerret egërsisht”.
-Metodi Çepreganov është specialist në neuropsikiatri dhe epileptologji me subspecializim në tomografi kompjuterike dhe rezonancë magnetike, dhe ka literaturë të përbashkët mjekësore me profesorë të shquar francezë. Për më shumë se 25 vjet ka ligjëruar si profesor në fushën e neurologjisë dhe psikiatrisë.
-Tezat e tij i sqaron në mënyrë shumë të thjeshtë, pa terme profesionale dhe fjalë të paqarta, me qëllim që ti kuptojë publiku i gjerë. Përpos se është një ekspert shumë i njohur edhe për tema aktuale dhe tërheqëse në veçanti, edhe pse është në dekadën e nëntë të jetës, Çepreganov akoma është mjek aktiv, mban ligjërata të famshme për stresin dhe sëmundjet tjera të kohës së sotme.
-Me profesorin Çepreganov folëm për atë se sa dhe si stresi ndikon mbi shëndetin tonë, si reagon organizmi më së shpeshti në gjendje stresuese, cilat janë pasojat dhe si mund të merremi me situatat stresuese. Çepreganov na sqaroi edhe për çrregullimet psikike, agresivitetin, krenarinë, zilinë dhe narcizmin. Ka shumë interes për ligjëratat e tij për stresin dhe depresionin.
-
-Si do i përkufizonit shtytjet destruktive dhe stresin, si probleme me të cilin ballafaqohet një përqindje e lartë e popullsisë?
-
--Shtytja destruktive është faktor endogjen, i trashëguar gjenetikisht, të cilin, me një kontroll të mirë mund ta sjellim në stabilitet racional, ndërsa stresi është faktor psiko-traumatik, i cili e sulmon integritetin e personit dhe i çrregullon parametrat e ekzistencës normale dhe çon në sëmundje psikosomatike.
-
-Çfarë do të thotë akumulimi i faktorit stresogjen dhe a ka zgjidhje dhe përgjigje?
-
--Akumulimi i faktorëve stresogjen janë psiko-trauma ekzogjene të cilat bëjnë revolucion në metabolizmin biologjik të organizmit tonë dhe për fat të keq secila gjendje stresogjene e zgjatur çon deri në gjenezë të sëmundjeve psikosomatike. Përgjigjja qëndron te ne, në tiparet tona të personalitetit dhe në aftësinë tonë për ta kontrolluar stresin dhe destruktivitetin që e mbartim brenda nesh, në emër të pasionit për grykësi për para, fuqi dhe famë.
-
-Çfarë thotë neuroshkenca lidhur me agresionin, arrogancën dhe dhunën?
-
-– Kjo ndodh nëse një person ka tipare të personalitetit, pa aftësi për kontroll adekuat të emocioneve, me pafuqi intelektuale, me papjekuri, me injorancë dhe privitizëm. Për shembull, edhe korrupsioni dhe ryshfeti janë pafuqi intelektuale të një mendjeje të robëruar dhe përfituese me mungesë të normave morale, sociale, shpirtërore dhe etike. Agresioni si pafuqi intelektuale për t’u marrë me bashkëbiseduesin është mungesë njohurish, mungesë e kapacitetit mendor dhe përqendrimit, e kulturës së komunikimit dhe qëndrimit ndaj dialogut në marrëdhëniet ndërnjerëzore.
-Në trurin tonë ka dy struktura kundërshtuese. Njëra është inteligjenca jonë emocionale, emocionet dhe jeta emocionale te secili prej nesh, ndërsa tjetra i kontrollon emocionet tona, e çon vendimin tonë në një proces mendor, se me çfarë fjalori, por edhe si dhe në çfarë mënyre, me çfarë gjestesh mimike, grimasë, ngjyrë të zërit dhe lëvizje do të jetë qëndrimi ynë ndaj bashkëbiseduesit, në shtëpi në rrugë, në punë etj.
-
-Cilin profil të personalitetit e konsideroni si më problematik?
-
--Nuk ka gjë më të keqe se vetë-pëlqimi, krenaria, zilia apo, me një fjalë, narcizmi dhe paranoja si mendime të gabuara. Narcizmi nënkupton shtytje destruktive njerëzore, hipertrofi e egos (me kryeneçësi, mungesë tolerance dhe inat), plotësimi i kënaqësisë nga kryeneçësia, mahnitje egoiste me atributet personale fizike dhe mendore që manifestohet me arrogancë dhe krenari. Personi i këtillë jeton për konflikte dhe ushqehet me to. Situatat konfliktuoze që i krijon vetë dhe janë nën regjinë e tij, i ka ushqim të patjetërsueshëm dhe shpirtëror, me mendje të robëruar dhe shumë shpesh vuajnë më shumë persona që ndodhen pranë tij. Asnjëherë nuk tregon dëshirë dhe prirje për mirëkuptim në marrëdhëniet ndërnjerëzore, me respekt dhe vëmendje të ndërsjellë, nuk e ka asnjëherë ndjenjën e fajit, përkundrazi – në çdo situatë konflikti gjithnjë di ta fajësojë palën tjetër dhe të hakmerret egërsisht.
-Megjithatë, më i rrezikshëm është narcizmi grupor i cili krijon një dukuri shumë të rrezikshme të solidaritetit dhe kohezionit të një popullate të caktuar. Grupi i pjekur është i lehtë për tu manipuluar dhe keqpërdorur, asgjë nuk mund ta ndalë, posaçërisht kur paraqitet gradimi, të qenit fanatik dhe fanatizmi. Arsyeja thjesht zhduket. Ka shumë shembuj në historinë botërore që ilustrojnë veprimet e histerisë masive si luftëra, greva, protesta, bërje tifo me tërbim nëpër stadiume, etj.
-
-Prej çfarë varet psiko-profili i një kombi?
-
--Nga të gjithë arketipet e vetes kolektive, nga e cila më vonë do të varen kornizat mendore dhe intelektuale si dhe kapacitetet e kuantumit të përgjithshëm pozitiv apo negativ të atij kombi, popullate apo populli, sjelljes së tij, kapaciteteve intelektuale, dijes, aftësisë për punë, tolerancës dhe fleksibilitetit, largpamësisë dhe kreativitetit, me një fjalë, nivelit të ekzistencës së tij, skajshmërish primitive nga njëra anë dhe tejet intelektuale, praktike, humane dhe e qëndrueshme nga ana tjetër.
-
-A ka zgjidhje për destruktivitetin njerëzor?
-
-Në teori po, por është e mundur vetëm në korrelacion me harmoninë dhe vetëdijen e lartë të një profili psikik, të dizajnuar dhe të programuar mirë, në një shoqëri të pjekur me norma të larta arsimore, edukative, shoqërore, juridike, ekonomike, morale, etike dhe shpirtërore me një bagazh të pasur njohurish.
-
-Prej çfarë varet parandalimi i agresionit?
-
-Më së pari nga arsimi dhe edukimi, familja, vendi i punës, kushtet e jetesës, angazhimi, motivimi për punë dhe marrja me sport si një faktor mjaft i rëndësishëm.
