--- content/portfolio_al/kur-je-ne-karroce-gjerat-qe-ne-maqedoni-jane-luks-ne-gjermani-jane-normale.md
@@ -1,48 +1,0 @@
----
-title: Kur je në karrocë “gjërat që në Maqedoni janë luks, në Gjermani janë normale”
-cover: '/images/uploads/kur-je-ne-karroce-gjerat-qe-ne-maqedoni.jpg'
-date: 2020-02-26
-author: ['Rita Behadini', 'Dajana Lazarevska']
-published:
- Diversity Media: 'http://diversitymedia.mk/kur-je-ne-karroce-gjerat-qe-ne-maqedoni-jane-luks-ne-gjermani-jane-normale/'
-
----
-
-Shkruajnë: Rita Behadini dhe Dajana Llazarevska
-
-Para do kohësh pashë një vido të disa aktivistëve në Tetovë, të cilët treguan se sa e vështirë është të kesh nevijë të ecësh me karrocë. Kjo më bëri të mendoj për gjithë atë vështirësitë që këta njerëz kanë në përditshmërë, kurse nga ana tjetër veç me pak më shumë kujdes mund ti japin aq shumë shoqërisë.
-Një i tillë është Gadaf Mememdi. Gadafi është studenti i parë me aftësi të veçanta që ka shfrytëzuar mundësinë e programit Erasmus për shkëmbim studentor jashtë vendit. Ai është student në Universitetin e Tetovës në degën gjuhë gjermane dhe ka vendosur që shkëmbimin ta realizojë në qytetin Marburg të Gjermanisë. Shkaku i karrocës me të cilën duhet të ecë, në Maqedoni Gadafi të shumtën e rasteve ka vështirësi në qasjen deri te objektet shkollore, institucionet apo edhe trotuaret anash rrugës. Objektet, rrugët, mjetet e transportit nuk kanë rampa për personat me aftësi të veçanta. E sipas Gadafit, kjo që në Maqedoni konsiderohet luks, në Gjermani është normale, sidomos rampat transportuese dhe ajri i pastër. Ai thotë se studentët duhet të shfrytëzojnë programet studimore për shkëmbim dhe se nëse aplikojnë për të njëjtën, nuk humbasin asgjë nga vetja por vetëm përfitojnë diçka shtesë.
-
-Çfarë të bëri të aplikoje për të bërë shkëmbim studentor jashtë vendit?
-
-– Idenë e vullnetarizmit dhe përfshirjes në gjëra të reja e kam pasur që gjatë kohë. Në fakt e kam praktikuar nëpër shkëmbimet dhe trajnimet e organizatës “LOJA”. Në një mënyrë, nëpërmjet vullnetarëve të huaj që vinin te kjo organizatë, bota erdhi te ne në Tetovë. I vetmi dallim kësaj herë ishte se ne shkuam drejt botës.
-Si të ka ndryshuar jeta tani që ke ikur për studime jashtë vendit?
-– Shumë gjëra kanë ndryshuar. Kam njohur shumë njerëz nga e gjithë bota. Dhe sa më shumë që njohim njerëz e kuptojmë se sa e vogël është kjo botë.
-A vëren dallime në sistemin e arsimit atje dhe sa e vështirë ishte të përshtateshe në fillim?
-– Ka shumë dallime në gjithë procesin arsimor këtu por ajo më e rëndësishmja do kisha thënë se është fakti që ka më tepër dedikim. Ekziston ajo ndjenja që gjithçka këtu është në vendin e duhur.
-Cilat janë dallimet më të mëdha nga ana e shërbimeve dhe sistemit jetësor që i vëren në krahasim me vendin tonë?
-– Gjithçka këtu është korrekte dhe funksionale, ashtu si duhet të jetë. Të kesh gjëra që funksionojnë si duhet në Maqedoni konsiderohet një luks. Këtu është thjeshtë një tjetër gjë normale. Dallimi është se te ne ende bëhet pazarllëk se a do të realizohet një e drejtë njerëzore apo jo. Këtu nuk është në pyetje a do të kemi rrugë ose jo, a do të kemi ajër të pastër ose jo. Këtu është në pyetje se në çfarë drejtimi do të zhvillohet vendi.
-
-A të është lehtësuar jeta atje, nga aspekti i qasjes nëpër shërbime publike dhe institucione?
-
-– Po më është lehtësuar sepse ashtu si thashë më lartë, gjithçka është në vendin e vet. Rampat në transportin publik dhe ajri i pastër nuk janë një luks për të cilin duhet të negociohet. Është një e drejtë që duhet t’i ofrohet çdo personi.
-
-Si është ndjenja të përjetosh kultura të reja?
-
-– Gjatë gjithë kësaj kohe që kam qëndruar këtu, e përjetova se si është të takosh kultura të reja, si është të përfshihesh në gjëra të cilat shumë njerëz mendojnë se nuk mund t’i bëjnë dhe kuptova se gjuha dhe kultura është një pjesë e esencës sonë. Kurse në thelb të gjithë jemi të njëjtë. Ne kemi gjuhë të ndryshme por ndjenjat tona janë të njëjta. Ndoshta flasim gjuhë të ndryshme por e qara dhe e qeshura janë të njëjta kudo në botë. Ndoshta kultura na dallon, por në esencë të gjithë jemi të njëjtë, të gjithë jemi njerëz.
-
-Sipas teje, përse studentët duhet të shfrytëzojnë ofertat për shkëmbime studentore?
-
-– Sepse është diçka që çdo student duhet ta shijojë dhe diçka që i ndryshon njerëzit për mirë. Është gjithmonë një ekperiencë që të ndryshon jetën pa marë parasysh ku shkon ose çka bën. Të jepet mundësia të shohësh një pjesë të re të botës. Nuk humbasim asgjë nga vetja, nëse shkojmë nëpër shkëmbime studentore vetëm do të përfitojmë diçka shtesë.
-
-Tani që ke jetuar edhe në Gjermani, a do të ktheheshe prapë të jetoje në Maqedoni?
-
-– Epo, disi duhet të kthehem se ashtu thotë në pasaportën time. Por kjo është pjesa legale. E para, është vendi ku unë jam lindur, dhe tjetra, është një vend ku unë kam jetuar shumë eksperienca. Nëse do më duhej të zgjedh, do të thoja të dyja. Do kisha dashur të sjell pjesë të Gjermanisë atje, por do kisha dashur edhe t’u jap atyre një pjesë tonën.
-
-Çfarë është diçka që të ka pëlqyer shumë atje dhe do doje të ishte njësoj edhe në Maqedoni?
-
-Vetëdija e njerëzve, fakti që secili e do vendin e vet dhe secili ka ndjenjën që duhet të ndjek një rregull thjeshtë sepse egziston dhe jo sepse do ta dënojë dikush. Këtu kam parë njerëz që presin në semafor në tre të mëngjesit, edhe nëse nuk ka asnjë veturë.
-
-– Çfarë mesazhi do ju jepje studentëve?
-Bëjeni. Çdo gjë që mendoni se nuk mund ta bëni, bëjeni. Nuk duhet të jetë pyetja a do të mundem, por si? E po erdhi një pikë ku do të mendoni se s’mundeni ta bëni vetë, atëherë kujtoni se nuk jeni vetë por ka njerëz rreth jush. Problemi mund të jetë një, por njerëzit rreth jush me miliona dhe metodat për të zgjidhur një problem, janë poashtu shumë më shumë se sa problemi. Pra, vazhdoni përpara, mendoni se çfarë mund të bëni vetë, pastaj në fakt do ta kuptoni se nuk jeni vetëm.
-Овој текст е продукт од проектот “Young Journalists’ Network” имплементиран од Демокраси Лаб. Овој проект е поддржан од Амбасадата на САД. Мислењата, откритијата и заклучоците или препораките изнесени овде се на имплементаторите/ авторите, и не ги одразуваат оние на Владата на САД.
