--- content/portfolio_al/me-nenvleresim-te-bulizmit-deri-ne-degradim-te-karakterit-te-femijeve.md
@@ -1,44 +1,0 @@
----
-title: Me nënvlerësim të bulizmit deri në degradim të karakterit të fëmijëve
-cover: '/images/uploads/bullying.jpg'
-date: 2020-01-26
-author: ['Dorina Ajdini', 'Maja Terziova']
-published:
- Diversity Media: 'http://diversitymedia.mk/%d1%81%d0%be-%d0%bf%d0%be%d1%82%d1%86%d0%b5%d0%bd%d1%83%d0%b2%d0%b0%d1%9a%d0%b5-%d0%bd%d0%b0-%d0%b1%d1%83%d0%bb%d0%b8%d0%b7%d0%bc%d0%be%d1%82-%d0%b4%d0%be-%d0%b4%d0%b5%d0%b3%d1%80%d0%b0%d0%b4%d0%b8/'
-
----
-
-Dora Ajdini/Maja Terziova
-
-Përqindja e nxënësve të shkollave të mesme në Tetovë që kanë qenë viktima të bulizmit brenda shkollës, nga bashkëmoshatarët e tyre, profesorët, por edhe nga organet më të larta të arsimit, është befasuese. Këtë e tregon hulumtimi i kryetarit të Unionit të nxënësve të shkollave të mesme (UNSHM) në Tetovë, Blendi Hodai. Sipas tij, bulizmi është temë për të cilën duhet të flitet më shpesh, për arsye se nuk mund ta tejkalojmë nëse nuk flasim për të. Por në vend të përkrahjes, Hodai është befasuar për të keq me sjelljen e drejtorëve të shkollave, nga të cilët ka marrë kërcënime pas shpalljes së rezultateve të hulumtimit.
-
-![Bullying](/images/uploads/bul-1.jpg)
-
-Hulumtuesit që kanë kryer anketimin me më shumë se 1729 nxënës të moshës 15-19 vjeçare, kanë përjetuar edhe një befasi tjetër. Pritjet e tyre kanë qenë se bulizmi fizik do të jetë më prezent te nxënësit. Por, rezultatet e hulumtimit kanë treguar se 79% të të anketuarve mendojnë se i ashtuquajturi bulizëm i padukshëm apo psikik është më prezent në raportin mësimdhënës-nxënës, por edhe mes vetë nxënësve.
-
-“Një e dhënë mjaftë shqetësuese që doli nga hulumtimi është se 44% e nxënësve kishin qenë viktima të bulizmit. Ky numër ka gjasë të jetë më i madh sepse bulizmi po bëhet rutinë e përditshme në shkollat tona”, tha Blendi Hodai, kryetar i UNSHM.
-
-
-**Çfarë është bulizmi dhe si ta njohim?**
-
-Bulizmi paraqet sjellje ofenduese ndaj personave të caktuar, të cilat përsëriten, janë të dëmshme dhe agresive. Dallon sipas nivelit të dhunës, ndërsa mund të jenë fizike apo psikike. Tema e bulizmit është temë e cila diskutohet shumë pak te ne. Nuk i kushtohet rëndësi e veçantë dhe lehet pas dore, jo vetëm nga prindërit dhe mësimdhënësit, por edhe nga institucionet shtetërore.
-Po, si ta dimë se një fëmijë është viktimë e bulizmit? Shenjat më të shpeshta janë kur refuzon të shkojë në shkollë, me pretekstin “Nuk dua të shkoj në shkollë, më dhemb barku, nuk ndihem mirë!”. Psikologët paralajmërojnë – nëse vërejmë se fëmija jep shenja të këtilla dhe janë të shpeshta, rasti duhet të hetohet.
-
-![Bullying](/images/uploads/bul-2.jpg)
-
-Fitore Bajrami, profesoreshë në Universitetin e Tetovës, në fushën e psikologjisë, për bulizmin thotë se përkufizohet si sjellje e qëllimshme agresive.
-“Personi i cili ka për qëllim të ngacmojë një person tjetër, e bën këtë me vetëdije të plotë, dëshiron ta dëmtojë tjetrin, duke filluar nga ajo “ta bëjë të ndihet keq”, dhe duke vazhduar në forma tjera. Ajo që është karakteristike te bulizmi është se personi që shkakton bulizëm, në këtë rast quhet agresor dhe gjithnjë ka më shumë fuqi se tjetri”, thotë profesoresha Bajrami.
-Një fokus i veçantë duhet t’i kushtohet edhe burimit të kësaj dukurie të paraqitur në shoqërinë tonë. Psikologja Edlira Destani thotë se duhet të ndiqet me kujdes secili ndryshim te nxënësit dhe të gjendet burimi i asaj sjelljeje.
-“Në momente të caktuara, kur nuk jemi në gjendje të shprehemi, zbatojmë asi mekanizma me të cilët bezdisim të tjerët për të ngritur vetveten. Kjo ndodh me të gjitha gjeneratat. Në ndërkohë, që të ngremë vetveten, fillojmë me maltretim verbal, tallje me fjalë, gjë me të cilën personi tjetër, që në këtë rast është viktima, nuk i kundërpërgjigjet ngacmimit, dhe duke e shmangur, kalon në fazë kur fillon të tërhiqet. Që ta shmangim bulizmin, duhet që secilin fëmijë që në moshë shumë të vogël, ta lejojmë të shprehet lidhur me atë që e shqetëson. Rasti duhet gjithnjë të merret seriozisht dhe duhet gjetur burimi se pse ai person sillet ashtu”, thotë Destani.
-Për të zgjidhur problemin mes moshatarëve në shkollë, fëmijët nuk duhet të kenë turp të flasin për ndodhinë me prindërit e tyre, të marrin më shumë përkrahje nga ata dhe angazhim më të madh nga psikologët. Theks i veçantë dhe kujdes t’i kushtohet shëndetit mendor. Gjithkush duhet të përpiqet për vetëbesim më të lartë, të mos lejojnë që dikush të ngrejë veten duke i nënçmuar ata. Në vend që të kritikojmë, të inkurajojmë.
-Por, sa janë profesorët të njoftuar me këtë dukuri dhe si reagojnë në rast se shohin një nxënës që ngacmon ose torturon një bashkëmoshatar?
-UNSHM drejton gishtin ka psikologët e shkollave të cilët, sipas atyre, duhet të mbajnë përgjegjësi për këtë gjendje, sepse nga ana e tyre nuk është kryer asnjë hulumtim. Nga Unioni thuan se psikologët as edhe që janë të njoftuar me nocionin “bulizëm”. Ata këtë e quajnë mosfunksionim i zyrës së psikologëve të shkollave. Sipas Unionit, ironia më e madhe se nxënësit as që e dinë se kush janë psikologët në shkollat e tyre.
-
-![Bullying](/images/uploads/bul3.jpg)
-
-**Bulizmi – akoma tabu temë në shoqëri**
-
-Profesoresha e shkollës së mesme, Lindita Emërllahu, konfirmon se për bulizmin nuk flitet mes profesorëve nëpër shkolla, e cila poashtu vë në pah disa mangësi në marrëdhënien profesor – nxënës.
-“Ne, mësimdhënësit, këtë koncept e kuptojmë si shumë të rëndësishëm, duke nisur nga vetvetja. Detyra jonë kryesore si profesorë duhet të jetë përqendrimi në atë se si të fuqizojmë karakterin e nxënësve që ata vetë të dinë se si duhet të reagojnë në rast bulizmi. Shumë nxënësve u mungon vetëbesimi. Nuk më pëlqen kur nxënësit gjithnjë përgjigjen me PO për gjithçka që iu them. Duhet të ketë marrëdhënie më të afërt midis nxënësve dhe profesorëve, duhet të ketë afërsi njerëzore, e cila do të ndihmojë në eliminimin e bulizmit shoqëror.
-Profesoresha Emërllahu mendon se bulizmin mund ta luftojmë edhe me ndryshimin e mënyrës së mësimdhënies duke aplikuar mësimdhënien interaktive. Në këtë mënyre nxënësit do të kyçen në mësimdhënie dhe do të fillojnë të mendojnë se mendimi i tyre është i rëndësishëm.
-E gjithë kjo është gjithashtu pjesë e veprimtarisë arsimore-edukative, se çfarë fëmijësh do të edukojmë në të ardhmen dhe në çfarë individësh do të shndërrohen.
