--- content/portfolio_al/miroslav-draganov-si-te-vazhdojni-me-buzeqeshje.md
@@ -1,43 +1,0 @@
----
-title: 'Miroslav Draganov: Si të vazhdoni me buzëqeshje'
-cover: '/images/uploads/miroslav-draganov.jpg'
-date: 2020-01-26
-author: ['Ljubica Ivanova', 'Arta Qerimi']
-published:
- Diversity Media: 'http://diversitymedia.mk/miroslav-draganov-si-te-vazhdojni-me-buzeqeshje/'
----
-
-Shkruajnë: Ljubica Ivanova dhe Arta Qerimi
-
-Ndan përvoja personale për të inkurajuar njerëzit tjerë dhe për të përcjellë se është më mirë të ndërmerret diçka si parandalim dhe të punohet në zhvillimin personal kur ekziston një themel i shëndoshë – Mirosllav Draganov, folës motivues dhe trajnues për vetëbesim, ish drejtor ekzekutiv i Institutit për të drejtat e njeriut, i cili ka pushtuar botën që në moshën 21 vjeçare me gara të të folurit publik, në të njëjtin rrafsh me udhëheqës botërorë. Është magjistër i së drejtës kushtetuese në Fakultetin e Drejtësisë në Shkup dhe është autor i librave “Në lëkurën tënde – vazhdo me buzëqeshje” dhe “Vazhdo me buzëqeshje – duke krijuar për dashurinë” që janë autobiografike dhe motivuese.
-
-Përpos librave të botuar, Draganov gjithashtu ndan përvoja dhe publikon video në rrjetet sociale, mban punëtori për zhvillim personal dhe vetëbesim, punon me individë, kompani, ekipe. Edhe audio versioni i librit të tij është në përgatitje e sipër, i cili nuk do të jetë për qëllime komerciale, ndërsa u dedikohet personave të verbër.
-“Bota e jashtme është një gjë, por si ne reagojmë ndaj botës së jashtme është plotësisht tjetër gjë, ajo që më së paku dëshiroj është që një tjetër të kalojë atë që kam kaluar unë”, thotë Mirosllav Draganov.
-
-Në librin e tij të parë rrëfen momente personale dhe tragjike kur, siç thotë ai, jo vetëm që u duhej një motivues, por ka qenë edhe në situatë për të mbërritur deri te psikiatri. Ai thotë se jeta praktikisht e ka plasur për toke.
-Me të folëm për këto momente kthyese, kyçe.
-
-Çfarë është ajo që ju jep fuqi të vazhdoni më tutje nëse e keni pasur vështirë?
--Jeta asnjëherë nuk do t’ju jetë e përsosur, por është i rëndësishëm përqendrimi, ndërsa ju zgjedhni se ku do të përqendroheni, thotë Mirosllavi dhe përkujton situatën kur vetëm pse ka besuar në fjalën e dhënë, ka ardhur në situatë të rëndë jetësore, megjithatë situatë prej së cilës ka mësuar mjaft.
--Ndoshta jam një nga të rrallët i cili për katër vite ka dhënë pesë dorëheqje, në kohë kur thuan se është vështirë të gjendet punë – thotë ai dhe rrëfen se si pasi që i është premtuat vend pune, arsye për të cilën ka lënë punë të sigurte, punëdhënësi nuk e ka mbajtur fjalën, ndërsa Mirosllavi ka ngelur pa punë.
-
--Motoja e jetës sime është se ka zgjidhje për gjithçka, përpos për ringjalljen e njeriut të vdekur”
--Kjo ishte periudhë e vështirë pasi as babai as njerka ime nuk punonin, vendosa mos t’ju tregoj asgjë, për dy javë mendonin se shkoj në punë, dilja i rregulluar nga shtëpia, e dija që nuk mund të më ndihmojnë dhe vetëm do të nervozojnë, ndërsa e dija që do të gjej zgjidhje, e mandej do ti njoftoj. Nga këndvështrimi i sotëm jam i kënaqur edhe kjo përvojë e kaluar, sepse ai i mbylli zyrat, ndërsa unë jam atje ku jam dhe akoma rritem, sqaron Draganov.
-
-Nga një jurist deri te një folës motivues. A gjeni lidhshmëri mes këtyre dy profesioneve, jo të gjithë mund të jenë një folës motivues, kur fillimisht keni ndier se ky mund të jetë profesioni juaj?
-
--Kur munden ata, pse jo edhe unë. U përqendrova plotësisht në punën si trajnues për vetëbesim dhe folës motivues, kuptova se kjo më bën më të lumtur, sado që duket e pasigurt në sytë e tjerëve. E punoj zyrtarisht atë që dua, edhe pse e kam bërë me vite, pa pagesë, privat, me njerëzit e mi të afërt, të cilët në vazhdimësi më kanë thënë se i ndikoj pozitivisht dhe më kanë nxitur. Drejtësia nuk është tërësisht e lënë pas dore në punën e tanishme sepse ky është parandalim, të ndërtoni marrëdhënie me njerëzit, me të punësuarit, të edukoheni, të dini të drejtat e juaja njerëzore që të mos lejoni dikush t’ua shkelë. Me tjerash, edukimi pozitiv në një ekip çon në produktivitet më të lartë, ai negativ çon në mobing, që është shkelje e të drejtave njerëzore, konsideron Mirosllav Draganov.
-
-Draganov thotë se ka ndjekur numër të madh të fjalimeve motivuese të huaja, por ka përjetuar edhe mjaft sfida në jetë, konsideron se mund ta zbatojë atë që e ka mësuar, duke shtuar elementin e vlerës së shtuar ashtu që kombinon teorinë dhe praktikën. Për vete thotë se i ka kaluar të gjitha shkallët, ka filluar prej më poshtë – praktikant, dhe deri më lartë, drejtor ekzekutiv.
-Thotë se reagimet e njerëzve e gëzojnë më së shumti. E thekson shembullin e magjistarit, Kristijan Shopov, i cili kryesisht ka punuar me fëmijë, por falë Mirosllavit, e ka plotësuar pasionin jetësor të punojë me të rritur, duke e tejkaluar frikën nga paraqitjet publike përpara publikut të gjerë, ndërsa Draganov është më së shumti është falënderues kur përvojat e tilla i shpërndajnë dhe përcjellin te persona tjerë.
-
-A mund çdo person të ketë vetëbesim? Si e keni ndërtuar vetëbesimin ndër vite, çfarë paraqet vetëbesimi për ju?
--Besoni në vete dhe në aftësitë tuaja, por mos e mbivlerësoni veten. Ka dy skajshmëri. Njëri moment është të qenit i vet-begenisur dhe narcisist, dhe tjetri është të mos besosh në vete me atë që nuk do ti njohësh aftësitë tua. Vetëbesimi është në mes dhe është më i bukuri, dhe shton se është në proporcion me emocionet nga skena, kështu që kur kemi më shumë emocione, kemi më pak vetëbesim, këshillon Draganov.
-
-Ai tregon për kolegen e tij e cili kishte shikuar të gjitha paraqitjet në TED TALKS, por kur duhej të dilte për të folur në konferencë, ajo kishte emocione dhe nuk ishte e suksesshme, andaj Draganov thotë se një nga teknikat për ndërtimin e vetëbesimit është përmes ushtrimeve.
--Nëse keni një emocione nga diçka, bëjeni më shumë herë. Mësoni si ta përmirësoni, por jo vetëm në teori, këshillon Draganov. Dhe harroni egon tuaj, shton ai.
-
-Draganov na tregoi se shpesh i thuan se ai nuk mendon si njerëzit në Maqedoni, kështu që e pyetëm se informatë kthyese ka te ne për atë që po bën.
--Jo gjithmonë na mungon cilësia në Maqedoni, por të jemi më të hapur për ndryshimet. Njerëzit këtu ndonjëherë dëshirojnë më parë të shohin a ka rezultate, thotë ai, dhe shton se kohët e fundit edhe ata që kanë qenë skeptikë, kur i ndiejnë rezultatet, rekomandojnë.
-Këshillë, si të vazhdojnë qytetarët me buzëqeshje, në kushte të një bote të ashpër, jete të shpejtë dhe probleme?
--Koha është kohë, nuk mund ta ndryshoj, por mund ta ndryshoj sjelljen time. Së pari le të ndërrojnë prioritetet, Ajnshtajni thotë se nëse me një mënyrë të të menduarit keni krijuar problem, mos pritni që me të njëjtën ta zgjidhni. Vetëm mënyra e të menduarit nuk do t’ju sjellë shumë dobi nëse nuk shoqërohet me veprime. Unë jam ai që thotë provoni, gabimet janë përvojë, por nëse doni të mësoni, jini të hapur për njohuri që të mbindërtoni aftësitë. Dhe bëni atë që ju bën të lumtur, jo vetëm në punë por gjatë gjithë kohës. Dhe vazhdoni me një buzëqeshje, është porosia e Mirosllav Draganov, jurist dhe folës motivues.
