--- content/portfolio_al/perfaqesimi-gjinor-ne-vend-nen-standardet-nderkombetare.md
@@ -1,24 +1,0 @@
----
-title: Përfaqësimi gjinor në vend, nën standardet ndërkombëtare
-cover: '/images/uploads/women-rights.jpg'
-date: 2020-01-26
-author: ['Teuta Buchi', 'Renata Penchova']
-published:
- Diversity Media: 'http://diversitymedia.mk/%d1%80%d0%be%d0%b4%d0%be%d0%b2%d0%b0%d1%82%d0%b0-%d0%b7%d0%b0%d1%81%d1%82%d0%b0%d0%bf%d0%b5%d0%bd%d0%be%d1%81%d1%82-%d0%b2%d0%be-%d0%b7%d0%b5%d0%bc%d1%98%d0%b0%d1%82%d0%b0-%d0%bf%d0%be%d0%b4-%d0%bc/'
-
----
-
-Teuta Buçi dhe Renata Pençova
-
-Barazia gjinore dhe përfaqësimi adekuat gjinor nëpër institucionet publike, duket të mbetet një realitet jo i afërt i Qeverisë, që pretendon se ka ndezur të gjitha motorët e saj për harmonizimin e standardeve dhe vlerave me Bashkimin Evropian, që më pas të bëhet edhe anëtare e saj. Marrja e timonit qeveritar nga ana e një partie me orientim te majtë socialdemokrat, nga ajo që kemi arritur të analizojmë, nuk ka reflektuar në zbehjen e dallimeve gjinore.
-Në krye të ndërmarrjeve publike, vazhdojnë të emërohen burrat dhe pozicionet udhëheqëse janë më të paarritshme për gratë. Sipas të dhënave që kemi siguruar nga Avokati i Popullit, shumica e ndërmarrjeve publike nuk kanë përfaqësim adekuat gjinor. Në tre ndërmarrjet që kemi analizuar, vërehet se burrat dominojnë në funksionet e emëruara dhe pozitat udhëheqëse. Në Radiotelevizionin Maqedonas (MRTV), të dy pozicionet e emëruara mbahen nga meshkuj. Nga 117 pozicione udhëheqëse, 65 janë gra. Nga numri i përgjithshëm i 794 të punësuarve, 55.29% janë burra, ndërsa 44.71% janë gra. Në ndërmarrjen publike për rrugë shtetërore, drejtues është emëruar person i gjinisë mashkullore, në pozicionet udhëheqëse prej 92 vendeve të punës, 74 apo 80.43% janë burra, ndërsa 18 apo 19,57% janë gra. Numri i përgjithshëm i të punësuarve në këtë ndërmarrje është 339, prej të cilëve 80,83% janë burra. Gjithashtu, Ndërmarrja Publike e Hekurudhave të Maqedonisë – Infrastruktura, nuk ka përfaqësim adekuat gjinor. Ndërmarrja udhëhiqet nga një mashkull. Në pozita udhëheqëse, nga 62 vende, 72.58% janë meshkuj. Sa i përket numrit të përgjithshëm të të punësuarve, prej 984, 909 ose 92.38% janë meshkuj.
-Nga NP Rrugë Shtetërore, duke sqaruar arsyen se pse është kështu gjendja e përfaqësimit gjinor, thonë se në konkurset e tyre paraqiten kryesisht meshkuj, ndërkohë që, sipas tyre, mundohen që të kenë më tepër kujdes në pozicionet udhëheqëse, ku punësojnë edhe gra.
-
-“Në Ndërmarrjen Publike për Rrugë Shtetërore numri më i madh i të punësuarve kryejnë detyra në pikat pagesore në territorin e Republikës së Maqedonisë së Veriut. Duke pasur parasysh kushtet për punë (punë në turne, turne të natës…) gjatë plotësimit të këtyre vendeve të punës ka më shumë interesim nga ana e meshkujve. Në drejtorinë e Ndërmarrjes Publike për Rrugë Shtetërore në shërbimet profesionale, numër i madh i të punësuarve në pozicionet udhëheqëse janë të plotësuara nga ana e grave”, thotë udhëheqësja e Repartit për resurse njerëzore dhe çështje të përgjithshme në këtë ndërmarrje, Biljana Cvetkoviq.
-Komiteti i Helsinkit për të drejtat njerëzore, vlerësojnë se përfaqësimi i grave në pozita udhëheqëse në sektorin shtetëror, por edhe atë privat, është shumë i ulët dhe lë shumë për të dëshiruar. “Duke pasur parasysh se vetëm 15% e pozicioneve ministrore u takojnë grave dhe se prej 80 prefektëve vetëm 6 janë gra, vijmë në përfundim se pjesëmarrja e grave në jetën politike është në një nivel shumë të ulët. Edhe pse nuk kemi të dhëna të sakta se sa është përfaqësimi i grave në institucionet shtetërore, përvojat tregojnë se edhe në sektorin shtetëror edhe në atë privat barazia gjinore dhe përfaqësimi i barabartë i grave nuk është realizuar. Aty ku ka përafrim të përfaqësimit të barabartë gjinor, vërehet se në pozicionet më të larta të punës paraqitet mospërputhje e madhe në favor të burrave, përderisa përfaqësimi i grave është më i lartë te pozicionet administrative”, kanë thënë nga Komiteti Helsinkit
-
-Edhe pse, numri i pjesëtareve të gjinisë femërore që kanë diplomuar në fakultete të ndryshme është i lartë, shifrat reduktohen dukshëm kur bëhet fjalë për materializimin e diplomave në vende pune.
-Sipas Entit Shtetëror të Statistikës, ku jepen të dhënat e studentëve të diplomuar në fakultetet e shkencave humane dhe ekzakte, në vitin 2018 kanë diplomuar 4531 studente, prej të cilave 3961 kanë ndjekur studimet në mënyrë të rregullt, ndërsa 570 kanë qenë me korrespodencë. Studente të diplomuara në institucionet shtetërore të arsimit të lartë janë 3895, ndërsa 566 në institucionet private. Numri më i madh i studenteve kanë diplomuar në Fakultetin Ekonomik, Fakultetin Filozofik dhe Fakultetin Filologjik. Ndërkaq, numri i studentëve që kanë diplomuar në vitin 2018 është 3167, prej të cilëve 2371 kanë qenë të rregullt ndërsa, 436 me korrespodencë.
-
-“Edhe pse të dhënat tregojnë se përqindje më e lartë e grave (59%) se sa e burrave (41%) përfundojnë arsim të lartë, më vonë gratë zhduken nga tregu i punës, përkatësisht punësimi i grave (34.6%) ulet në raport me punësimin e burrave (53.6%). Kjo tregon në atë se nuk bëhet fjalë për mungesë të grave të arsimuara dhe të kuadrove të gjinisë femërore, por për ekzistimin e barrierave gjatë hyrjes në tregun e punës si dhe për përparimin e mëtutjeshëm në karrierë. Kjo është pasojë e normave gjinore tradicionale shoqërore, të cilat akoma kanë ndikim ndaj pjesëmarrjes së gruas në tregun e punës, ndërsa duke pasur parasysh balansin që gruaja duhet të bëj mes jetës private dhe profesionale”, vlerësojnë nga Komiteti Helsinkit.
-“Të drejta të barabarta për burrat dhe gratë “ është një ndër parimet themelore të Kartës së Kombeve të Bashkuara të miratuar nga udhëheqësit botërorë më 1945. Parimi i barazisë gjinore u njoh edhe në Deklaratën Universale të të Drejtave të Njeriut në vitin 1948, e cila afirmoi “të drejtat e barabarta të burrave dhe grave”. Integrimi i perspektivës gjinore është procesi i vlerësimit të implikimeve për gratë dhe burrat për çdo veprim të planifikuar, përfshirë legjislacionin, politikat apo programet, në të gjitha fushat dhe në të gjitha nivelet.
