--- content/portfolio_al/shtohet-denoncimi-i-rasteve-te-dhunes-ne-familje.md
@@ -1,36 +1,0 @@
----
-title: Shtohet denoncimi i rasteve të dhunës në familje
-cover: '/images/uploads/shtohet-denoncimi-i-rasteve-ne-familje.jpeg'
-date: 2020-03-26
-author: ['Teuta Buchi', 'Evi Shkopi']
-published:
- Diversity Media: 'http://diversitymedia.mk/shtohet-denoncimi-i-rasteve-te-dhunes-ne-familje/'
-
----
-
-Shkruajnë: Evi Shkopi, Teuta Buçi
-
-Gjatë viteve të fundit është rritur përqindja e denoncimit të rasteve të dhunës në familje. Ana Avramoska – Nаushkova, ka mundësinë të trajtojë drejtpërdrejtë raste të tilla, pasi ajo është këshilltare ligjore tek Rrjeti Kombëtar Kundër Dhunës ndaj Grave dhe Dhunës në Familje. Këto raste jo rrallë herë ndikojnë emocionalisht tek ajo. Avramoska – Nаushkova thotë se më së shpeshti pre e dhunës janë femrat. Madje situata është edhe më e komplikuar kur ato janë të varura financiarisht nga dhunuesi, kjo vështirëson edhe më shumë largimin e tyre nga ambienti i dhunshëm. Në rastet e dhunës në familje, fëmijët veçanërisht kanë gjendje të rënduar psikologjike. Në vijim ju sjellim intervistën e plotë që realizuam me të:
-
-A keni shifra që tregojnë se si është gjendja në Maqedoni sa i takon dhunës në familje?
-
-– Kemi disa shifra, të cilat janë siguruar nga disa institucione zyrtare si nga Ministria për Punë të Brendshme. Në raportin vjetor për vitin 2018 thuhet se janë paraqitur 1006 raste të dhunës në familje. Nga Ministria për Punë dhe Politikë Sociale, për vitin 2018 si raste të reja të regjistruara janë paraqitur 919. Nëse bëhet një krahasim i raporteve të Ministrisë së Punëve të Brendshme për vitin 2017 dhe vitin 2018, rezulton se paraqitja e dhunës në familje është rritur për 11%.
-
-Kush rezulton të jetë më shpesh pre e dhunës në familje?
-
-– Pjesa më e madhe e këtyre paraqitjeve të dhunës në familje është bërë nga femra, të cilat më së shpeshti janë pre dhunën në familje dhe atë dhunë fizike në 85% të rasteve, nga partneri i tyre intim ose partneri në martesë. Në të dhënat e Ministrisë për Punë dhe Politikë Sociale për vitin 2018, pjesa më e madhe e rasteve janë paraqitur nga femrat, më pas nga meshkuj dhe nga fëmijë. Kemi një statistikë të vogël që arritëm ta sigurojmë në shtator, përmes ushtrimit të të drejtës për Qasje të Lirë deri tek Informacionet me Karakter Publik, pra kërkuam të dhëna nga Ministria për Punë dhe Politikë Sociale, për statistikë për dhunë për periudhën janar-qershor vitin 2019 është 767. Tani se a është rritur guximi për të paraqitur raste të tilla, apo a është rritur dhuna në familje, kësaj mund t’i japë përgjigje ndonjë studim tjetër.
-
-Çfarë ndërmerrni konkretisht ju si organizatë kur një femër e dhunuar ju drejtohet për ndihmë?
-Do t’u them një rast konkret. Një grua, nënë e dy fëmijëve të mitur, viktimë e dhunës në familje, u drejtua tek ne. Pas diskutimit me këshilltarin ligjor, u arrit në përfundimin se mund të merren më shumë veprime juridike që gruaja të fuqizohet financiarisht, në përputhje me Ligjin për Mbrojtjen Sociale dhe Ligjin për Mbrojtjen e Fëmijëve. Gjithashtu gjatë asaj periudhe femra e cila ishte pre e dhunës në familje kishte nevojë për ndihmë urgjente për ushqim dhe higjienë, iu sigurua paketë me fonde themelore, përmes Fondit Ndërhyrës të Grave. Në bashkëpunim me organizata të tjera, zonja në fjalë mori edhe veshmbathje për të dhe dy fëmijët e saj të mitur.
-
-A kanë frikë femrat që të paraqesin dhunën që ushtrohet ndaj tyre?
-
-– Sipas hulumtimit të fundit të OSBE-së, për vitin 2018 që përfshiu të gjithë Ballkanin Perëndimorë, pra edhe Republikën e Maqedonisë së Veriut, rezulton se 42% e femrave që janë pre e dhunës në familje, për shkak së nuk kanë mundësi financiare, mospasjes së pronës apo mungesës së mbështetjes nga të afërmit ose nuk largohen nga ambienti ku janë pre e dhunës, ose largohen dhe më pas kthehen përsëri. Gjithashtu në hezitimin për paraqitjen e dhunës ndikon edhe paragjykimi i opinionit, fjalët e njerëzve mes njeri-tjetrit, i tregojnë me gisht, në ambientet e vogla është edhe më e theksuar kjo gjë, sepse të gjithë njihen me njëri -tjetrin. Shqetësues është niveli i lartë mungesës së besimit në institucione, që jo gjithmonë është i arsyetuar. Mendoj se problem kryesorë i viktimave që janë pre e dhunës në familje, është së pari gjendja e tyre psikologjike, më pas gjendja financiare dhe ndihma juridike që duhet t’u jepet. Sepse nëse e hapin vet procedurën për këtë sigurisht që ajo ka kosto materiale. Megjithatë nëse Qendra për Punë Sociale konstaton rast të dhunës në familje, në bazë të detyrës zyrtare duhet të nisë procedurë për shqiptim të masave të nevojshme.
-
-Në bazë të informacioneve që ju keni, a veprojnë në kohë institucionet kompetente kur paraqitet ndonjë rast i dhunës në familje?
-
-– Grupet ku ne jemi fokusuar në përputhje me analizën që ne kemi bërë me femrat që janë pre e dhunës në familje, treguan ankesa apo njëfarë mosinteresimi të Qendrave për Punë Sociale. Në princip kur kemi një mekanizëm mbrojtës që duhet të vendoset në dispozicion të viktimave që janë pre e dhunës në familje, rolin kryesorë e ka Qendra për Punë Sociale, më pas policia dhe pastaj punonjësit shëndetësorë.
-
-A keni informacione se vallë a janë ndërtuar në Maqedoninë e Veriut kapacitetet e nevojshme për të mbrojtur personat që janë pre e dhunës në familje?
-
-– Ministria për Punë dhe Politikë Sociale ka treguar publikisht se sa kapacitete ka gjithsej, në përputhje me Konventën e Stambollit që tek ne filloi të zbatohet në kuadër ta planit aksional që filloi nga Tetori i vitit 2018. Kemi kapacitete mjaft të kufizuara. Qendra strehuese që gjenden në 5 qytete në Maqedoni, ku kapaciteti gjithsej është 35 krevate. Kjo, pra, në nivel nacional. Konsiderojmë se nuk është e mjaftueshme dhe se duhet të bëhet përputhje me Konventën e Stambollit, të bëhet një përshtatje dhe të plotësohen standardet.
