--- content/portfolio_al/ushqimi-organik-ne-vend-akoma-ne-fillesen-e-tij.md
@@ -1,55 +1,0 @@
----
-title: Ushqimi organik në vend – akoma në fillesën e tij
-cover: '/images/uploads/organic-food.jpg'
-date: 2020-01-10
-author: ['Ljubica Ivanova', 'Arta Qerimi']
-published:
- Pina: 'https://pina.mk/2959-organska-hrana-vo-zemjava-se-ushte-vo-zachetok/'
- Diversity Media: 'http://diversitymedia.mk/%d0%be%d1%80%d0%b3%d0%b0%d0%bd%d1%81%d0%ba%d0%b0%d1%82%d0%b0-%d1%85%d1%80%d0%b0%d0%bd%d0%b0-%d0%b2%d0%be-%d0%b7%d0%b5%d0%bc%d1%98%d0%b0%d0%b2%d0%b0-%d1%81%d0%b5-%d1%83%d1%88%d1%82/'
-
----
-
-**Qytetarët nuk e ndjekin shumë trendin botëror të përdorimit të ushqimit organik. Pjesërisht, kjo u përshkruhet çmimeve më të larta të prodhimeve organike, por edhe faktit se shitoret ku mund të gjendet ushqim organik janë të rralla. Përkundër rekomandimeve të përditshme të nutricionistëve për konsumim të rregullt të prodhimeve organike, mbetet e hapur çështja se sa këtë ushqim mund ta blejnë qytetarët.**
-
-Ljubica Ivanova dhe Arta Qerimi
-
-Shkupjanët janë konsumuesit më të mëdhenj të ushqimit organik i cili vjen nga i tërë vendi, thuan prodhuesit. Një nga prodhuesit e ushqimit organik është Angelço Dimitriev nga Shtipi, i cili, bashkë me familjen e tij, gati dhjetë vite merret me prodhimtari organike, ndërsa prej para pak kohësh edhe me përpunimin e qershive, pjeshkave, kajsive, mollëve, dardhave, bajameve, lajthive. Kanë filluar me plantacione më të vogla bujqësore, ndërsa tani veç më punojnë në sipërfaqe prej 25 hektarësh në Zarapinci të Probishtipit, një pronë mes Sveti Nikoles dhe Probishtipit, në një mjedis të pastër, të pa ndotur.
-
-“Duhet pak kohë më tepër për prodhimtarinë organike në krahasim me atë konvencionale.
-Prodhimet tona i distribuojmë në shitore për ushqim të shëndetshëm, organik në Shkup. Për çmime nuk bëjnë problem, i blejnë menjëherë prodhimet organike. Për dallim prej Shkupit, ku ka kërkesë më të lartë, në qytetet tjera në Maqedonisë, interesimi është më i ulët. Akoma nuk është zhvilluar vetëdija për ushqimin organik, njerëzit dyshojnë, andaj ka nevojë për më tepër promovim, por ka shpresë”, thotë Angelço Dimitriev – bujk.
-Familja Dimitriev, në bashkëpunim me një agjenci turistike nga Shkupi, planifikon të organizojë vizita në plantacionet e tyre organike, me qëllim të zhvillimit të turizmit rural dhe promovimit të ushqimit organik, në mënyrë që njerëzit të sigurohen për cilësinë e prodhimit. Ata posedojnë edhe certifikatë nga kompania “Balkan Biosert”, e cila kryen kontrollin e prodhimtarisë së tyre, gjë për të cilën paguajnë nga 40.000 denarë për një kontroll, ndërsa shteti iu kthen prodhuesve 50% të shumës.
-
-Për dallim nga trendët botërore, vendi jonë pothuajse nuk ekziston në statistikat botërore, të cilat përfshijnë shtete me prodhimtari dhe kërkesë të përparuar të ushqimit organik. Konsumi i prodhimeve organike nuk e ndjek trendin e shtetit, i cili e stimulon prodhimtarinë organike përmes subvencioneve shtetërore, që janë më të larta për 30% në krahasim me atë konvencionale. Përmes Agjencisë për përkrahje financiare të bujqësisë dhe zhvillimit rural, në tetor të vitit 2019, në llogaritë e 638 kultivuesve të prodhimeve organike, janë paguar mbi 82 milion denarë (MBPEU).
-
-**Ushqim organik ka vetëm në vende të caktuara**
-
-Dekani i Fakultetit të bujqësisë në UGD – Shtip, prof.dr. Ljupço Mihajllov ndjek trendin e
-konsumit të prodhimeve organike te shitjet me pakicë.
-
-“Krejt rastësisht, në marketet më të mëdha, mund të vërehet prezenca e vendeve të caktuara për shitje të prodhimeve të etiketuar dhe certifikuara organike, por në vendin tonë, fatkeqësisht, akoma nuk mund të konstatohet një ofertë dhe kërkesë e qëndrueshme e as treg dhe bursë seriozisht e zhvilluar e ushqimit organik” sqaron profesori Mihajllov.
-Shumica e zinxhirëve të mëdhenj të marketeve nuk mbajnë prodhime bimore organike në të gjithë pikat shitëse të tyre. Se në cilin vend do të shiten prodhimet organike vendoset sipas frekuentimit dhe kërkesës së konsumatorëve.
-
-Sipas pronares së një shitoreje të ushqimit organik në qendër të Shkupit, për shkak të çmimeve më të larta të prodhimeve organike, kërkesa për këto prodhime në tregun vendor është e ulët.
-
-“Çmimet e produkteve organike janë ndonjëherë edhe dyfish më të larta se sa të produkteve inorganike, por kemi klientë që përdorin produkte organike në përditshmëri, mësohen me to dhe i blejnë pa marrë parasysh çmimin”, thotë pronarja e shitores organike.
-
-Tregu Evropian i prodhimeve organike u rrit për më shumë se 37 miliardë euro në vitin 2017, ndërsa secili evropian harxhon 67 euro për ushqim organik.
-
-
-**Prodhimtaria është akoma në fillesën e saj**
-
-“Dobra zemja” është një nga shitoret që blen ushqimin organik maqedonas lokal. Ivo Baru, i punësuar në shitore, thotë se shumica e prodhimeve sigurohen nga prodhues të fshatrave rreth Koçanit, Strumicës, Kavadarit, Negotinës, Resnjës, Vallandovës, si dhe nga fshatrat e Shkupit, nga prodhues që kanë certifikatë për prodhimtari organike, por që të binden për këtë, më së shpeshti edhe personalisht i vizitojnë prodhuesit para se të fillojnë bashkëpunimin me to. Sa u përket çmimeve, nuk ka ndonjë formulë të caktuar sipas së cilës i caktojnë çmimet. Sipas Barus, çmimet kryesisht varen nga ajo se sa është prezent produkti në treg, si dhe nga çmimi që kanë caktuar vetë prodhuesit.
-
-*“Prodhimtaria organike në Maqedoni sikurse është akoma në fillesën e saj. Numri më i madh i konsumatorëve më besnik në shitoren tonë janë kryesisht nëna, të cilat blejnë ushqim organik për fëmijët e tyre të vegjël. Përpos nënave, si konsumatorë të ushqimit organik paraqiten edhe njerëz me probleme shëndetësore, njerëz të cilët i përmbahen ndonjë regjimi të caktuar të të ushqyerit, që jetojnë shëndetshëm dhe të cilët janë në dietë. Shitorja ekziston tani më pesë vite dhe fatkeqësisht nuk mund të lavdërohemi me rritje të kërkesës siç i kemi ne pritjet”*, deklaroi Ivo Baru – i punësuar në shitoren për ushqim organik “Dobra zemja”.
-
-Kemi bërë një krahasim të çmimeve të frutave dhe perimeve të prodhuara në mënyrë organike dhe inorganike.
-
-Një kilogram mollë organike kushton 77 denarë, kilogrami i mollëve inorganike kushton 32
-denarë, marula organike është 37 denarë, ndërsa inorganike është 20 denarë, një copë
-patëllxhan organik kushton 35 denarë, ndërsa një kilogram inorganik është 59 denarë.
-Sipas Planit nacional për prodhimtari organike, në Maqedoni mbizotëron informim i
-pamjaftueshëm i konsumatorëve për përfitimet nga përdorimi i prodhimeve organike
-bujqësore, informim i pamjaftueshëm i tregtarëve për ekzistimin e subvencioneve për shitje të prodhimeve organike, treg vendor i pazhvilluar mjaftueshëm dhe mungesa e shitoreve të specializuara për ushqim organik, veçanërisht në vendbanimet më të vogla.
-
-Nuk ka aktivitete të mjaftueshme për të promovuar prodhimtarinë organike dhe për të ngritur vetëdijen e konsumatorëve, si dhe sasi të pamjaftueshme të produkteve organike për furnizim të vazhdueshëm të supermarketeve.
-Hulumtime të shumta tregojnë për efektet pozitive të ushqimit organik, si frutat dhe perimet që përmbajnë më shumë vitamina. Prodhimet organike përmbajnë më pak mbetje të metaleve të rënda, nitrateve dhe pesticideve, qëndrojnë më gjatë në depo, pa e humbur cilësinë e tyre dhe ndikojnë në mbrojtjen e ambientit natyror. Studime të shumta vërtetojnë se niveli i emetimit të dioksid karbonit për hektar mund të jetë 50% më i ulët në bujqësinë organike krahasuar me atë konvencionale.
