--- content/portfolio_mk/arsimtaret-ne-maqedoni-kane-rroga-me-te-uleta.md
@@ -1,50 +1,0 @@
----
-title: Наставниците во Македонија имаат најниски плати и највисоки животни трошоци
-cover: '/images/uploads/salary.jpg'
-date: 2020-01-26
-author: ['Rrita Syla', 'Ardit Ramadani']
-published:
- Diversity Media: 'http://diversitymedia.mk/%d0%bd%d0%b0%d1%81%d1%82%d0%b0%d0%b2%d0%bd%d0%b8%d1%86%d0%b8%d1%82%d0%b5-%d0%b2%d0%be-%d0%bc%d0%b0%d0%ba%d0%b5%d0%b4%d0%be%d0%bd%d0%b8%d1%98%d0%b0-%d0%b8%d0%bc%d0%b0%d0%b0%d1%82-%d0%bd%d0%b0%d1%98/'
-
----
-
-Пишуваат: Рита Сила и Ардит Рамадани
-
-На сметките на двајца родители во семејството врабoтени во основно образование во Македонија, на крај на месецот се исплаќаат вкупно 800 евра, во соседна Албанија 1,000 евра, додека во Косово 1,098 евра. Бројките се менуваат и стануваат алармантни во ситуација кога едниот од родителите работи за минимална плата. Наставниците се жалат дека не можат да го поминат месецот со овие плати, додека според експертите, во Македонија, освен што платите се пониски, трошоците за живот се повисоки во споредба со Албанија и Косово.
-
-*Македонија*
-
-Месечната плата на еден наставник од основно образование во Македонија е 400 евра, што значи дека едно семејство со двајца вработени како наставници во основно образование има месечен приход од 800 евра. За минималната потрошувачката кошничка на едно четиричлено семејство, во која влегуваат исхрана, пијалоци, сметки, хигиена, превоз, облека и одржување на здравје и култура, според Сојузот на синдикати на Македонија, се потребни 534 евра.
-Одземајќи ги основните месечни трошоци, на едно семејство со двајца вработени како наставници во основно образование на крај на месецот му преостануваат 266 евра кои не се неопходни основни животни трошоци.
-Ако ги собереме приходите на едниот родител вработен во образование со другиот, кој работи за минимална плата (235 евра), како *дополнителни средства на ова семејство му остануваат 106 евра.
-
-*Албанија*
-
-Месечната плата на наставник од основно образование во Република Албанија е 500 евра, и според ова, на сметка на семејство со двајца вработени наставници, на крај на месецот се исплаќаат 1000 евра, додека за месечните трошоци на едно четиричлено семејство, според Државниот завод за статистика на Албанија „Инстат“, се потребни 619 евра.
-Ако ги одземеме месечните трошоци од вкупната сума на приходи, за ова семејство како дополнителни средства му остануваат 381 евра.
-Од друга страна пак, кога се работи за четиричлено семејство, каде едниот родител е вработен во образование а другиот работи за минимална плата, како дополнителни средства му остануваат 92 евра.
-
-![Salary](/images/uploads/salary-mk.png)
-
-*Косово*
-
-Еден наставник во Република Косово секој месец зема плата од 549 евра. Така, на сметка на едно семејство со двајца вработени наставници во основно образование, се исплаќаат 1098 евра. Од друга страна, за едно семејство да го помине месецот, според косовската Агенција за статистика се потребни 650 евра.
-
-Освен потребните месечни трошоци за едно четиричлено семејство во Косово, им остануваат 448 евра за дополнителни трошоци.
-Доколку едниот од родителите работи за минимална плата, а другиот е вработен во образование, како дополнителни средства или средства кои не се неопходни, им остануваат 69 евра.
-
-![Salary](/images/uploads/salary-mk-2.png)
-
-Според наставничката Гонџе Мемеди, платата која ја зема не и е доволна за живот. Што се однесува до покачувањето на платите во образованието, смета дека ова мало покачување не е доволно.
-„Да бидеш наставник во Македонија не е воопшто лесно, почнувајќи од техничките проблеми до најниската плата во регионот, која не е доволна ни за покривање на месечните трошоци и ја загрозува иднината на оваа професија. Иако, во минатите години имало мали покачувања на платите, сепак не ги задоволуваат основните потреби споредено со високите цени во секоја економска сфера“, нагласи Гонџе Мехмеди.
-
-За експертот за статистика Шпреса Сила, податоците обезбедени од наша страна се загрижувачки, бидејќи, според неа, овие бројки влијаат само на демотивирање на наставниците. Сила го повикува Синдикатот за образование да инсистира на покачување на платите.
-„Ако се анализираат платите на семејствата кои имаат двајца членови наставници, во земјите во регионот, забележуваме дека во Македонија, освен што платите се пониски, од една страна, трошоците за живот се повисоки, од друга страна. Според изнесените податоци за плати, во Македоинија, 66.75%, 59% во Албанија и 61.9% во Косово, се наменети за месечната кошничка на трошоци,вели Сила и додава:
-
-*„Го охрабруваме синдикатот во образованието да инсистира на покачување на платите. Едно покачување на платите од 10% би ги зголемило средствата за дополнителни трошоци од 33.25% на 39.31%. Во спротивно, задолжителното средно образование, недостигот на можности за развој на кариерата на наставниците, ниските плати, би влијаеле врз мотивираноста на наставниците и континуираното зголемување на квалитетот на образованието“*, смета Шпреса Сила.
-
-Земајќи го сето ова во предвид, може да заклучиме дека едно семејство со една наставничка плата и со една минимална плата, ниту во Косово, ниту во Македонија, ниту во Албанија, можат да излезат на крај за други дополнителни трошоци, бидејќи, дополнителните средства кои им остануваат се многу мали, и тоа: Косово (69 евра), Албанија (92 евра), Македонија (106 евра).
-
-*Под поимот дополнителни средства или дополнителни трошоци ги подразбираме оние средства кои остануваат откако ќе се плататат основните животни трошоци. Тука спаѓаат одморите или други лични потреби кои не се вклучени во месечните трошоци.
-
-Извори: Косовска Агенција за Статистика, Сојуз на Синдикати на Македонија, Државен Завод за Статистика, Инстат, Еуростат.
