--- content/portfolio_mk/jovana-georgieva-mbajtese-medaljeje-ne-karate-sfida-me-e-madhe-eshte-qe-frika-te-shnderrohet-ne-liri.md
@@ -1,37 +1,0 @@
----
-title: 'Јована Георгиева,европски медалист во карате:Најголем предзвик е стравот да се претвори во слобода'
-cover: '/images/uploads/jovana.jpg'
-date: 2020-02-21
-author: ['Teodora Noveska', 'Shefkije Alasani']
-published:
- Diversity Media: 'http://diversitymedia.mk/%d1%98%d0%be%d0%b2%d0%b0%d0%bd%d0%b0-%d0%b3%d0%b5%d0%be%d1%80%d0%b3%d0%b8%d0%b5%d0%b2%d0%b0%d0%b5%d0%b2%d1%80%d0%be%d0%bf%d1%81%d0%ba%d0%b8-%d0%bc%d0%b5%d0%b4%d0%b0%d0%bb%d0%b8%d1%81%d1%82-%d0%b2/'
----
-
-Пишуваат: Теодора Новеска и Шефкије Аласани
-
-„Кога имав пет години татко ми ме однесе на карате, оттогаш таму останав. Немав типично детство, постојано бев на тренинг, верував во соништа, барав медали, се соочував со стравови од кога знам за себе “
-
-Насмеана, ведра и разговорлива, Јована Георгиева е државен сениорски првак и европски бронзен медалист во карате. Да се биде спортист според неа е голема привилегија, а каратето и’ претставува најголема љубов.
-
-Кога била дете била палава и хиперактивна, се интересирала за се и сешто, сакала да знае многу, а енергија вели имала колку што сега и недостасува.
-
-„Татко ми ме однесе на карате на мои 5 години. За рака! И така таму останав. 2020 година ми е полнолетството во карате, значи вежбам 18 години. Започнав во карате клуб „Астибо“, најголем дел поминав во „Македонка“, а сега тренирам во „Олимпик“. Немав типично детство. Постојано бев на тренинг. Верував во соништа, барав медали. Се соочував со стравови од кога знам за себе. Желбата за карате беше детска желба, ама со годините минуваше во некоја поразлична форма – надвладување над себе, соочување со самата себе, желба да се терам да одам напред. Каратето ми е најголема животна лекција. Љубовта и желбата водат напред, но патот го креира истрајноста и вербата во себе“, објаснува Јована.
-
-Седејќи спроти мене, ми дава впечаток на силна жена, но со грациозни потези. Лесно може да биде сечиј идол. Ја прашувам дали има свој идол или некој на кој се угледува. Вели, нема идол. Има жени на кои им се восхитува, но сфаќа дека самата станува таква и се гордее со тоа. Единствена инспирација и е мотивацијата да се надмине себеси и стравот да го претвори во слобода. Одговорот уште повеќе ми ја зацврстува сликата за неа и почнав да и’ се восхитувам.
-
-Државен сениорски првак е. Има европски бронзен медал во 2017 и балкански бронзен истата година.
-„Да се биде спортист е голема привилегија. Ама се стекнува како резултат на постојаната жртва која ја даваш за да бидеш во својата максимална физичка и психичка форма. Пречки има милион. Прва пречка ми е што денот е 24 часа. Секоја друга лекција од карате ми била само олеснување. Чувството да си европски медалист е нешто кое не може да се опише со збор. Замисли, со години ги гледаш тие ликови кои освојуваат. Фиктивни ти се. И одеднаш, после 10 години тренинг, без никаков медал и надеж, во мојата прва студентска година на факултет, јас освојувам балкански, па европски медал и сум меѓу нив. Чувството е неверојатно. Се’ за што си копнеел како дете, ти се остварува. Не си свесен дека тоа е тоа за кое си правел се’ што си правел секој ден. Секое станување во осум – добива смисла. Секој удар и секоја мисла која си ја посветил за сонот,“ вели Јована.
-
-Мене интересен ми беше податокот дека на Европското првенство ја победи дотогашната европска шампионка. Ја прашав како се чувствува во врска со тоа. Искуството, вели, било поинакво, оти не верувала во себе толку, колку што била препуштена на моментот и концентрацијата и’ била огромна. Сакала да се изненади самата себе и успеала.
-
-Четврта година е, студира молекуларна биологија на ПМФ. Факултет, кој според најголем број од луѓето кои ќе ги прашаш, е тежок и бара многу внимание. Не знам дали нејзе и’ е така, но со леснотија и голем ентузијазам зборува за факултетското искуство.
-„Факултетот ми е љубов. Како и се’ друго што работам. Затоа што, се’ што правам го правам со љубов. Многу ми е тешко да живеам во ден со 24 часа, затоа што сакам да стигнам и да средам се’. Да учам, да вежбам, да читам и да спијам. Социјалниот живот ми страда многу, ама во последно време сфаќам дека не губам многу, не додека се’ тоа што го правам ме наградува и ме полни со енергија. Ја сакам природата и колку различните форми на живот се совршени, како енергијата нашла тек со материјата. Факултетот ме спушта на земја. Не е лесно да се комбинира со тренинг, ама на крајот на денот, кога знаеш колку правиш за себе, тоа те прави среќен“, објаснува Јована.
-
-Зборувајќи за иднината, вели оти не знае каде се гледа. Има многу интереси и идеи, знае само дека сака да живее со предизвици. Да биде во тек со случувањата, да го пренесе тоа што го знае и гради изминативе години. Исполнета се чувствува кога се победува себе си, оти е во трка со никој друг, туку со самата себе.
-
-Завршуваме со разговор за каратето. Ја прашувам каква е ситуацијата во Македонија.
-„Македонија има одлично карате, многу конкурентно на големите сили. Ние веќе имаме светски имиња. Да тргнеме од Павлов во сениорска, па младинските медали, олимпискиот медал на Весели, европските медали на Смилковски, Свинарски. Светскиот на Бојан Костов. Младите Милева и Божиновска со балкански медали. Девојка за почит што спојува медицина и карате, европската шампионка Вероника Костовска. Македонија има исклучителни таленти, што ми е мило да ги имам покрај мене“, го завршува разговорот Јована Георгиева.
-
-Се поздравуваме со Јована а во глава ми одзвонува нејзиното мото: „Мостовите што ќе ги изгорам нека ми го осветлат патот“.
-
-Овој текст е продукт од проектот “Young Journalists’ Network” имплементиран од Демокраси Лаб. Овој проект е поддржан од Амбасадата на САД. Мислењата, откритијата и заклучоците или препораките изнесени овде се на имплементаторите/ авторите, и не ги одразуваат оние на Владата на САД.
