--- content/portfolio_mk/korrupsioni-dhe-ryshfeti-eshte-nje-shprehje.md
@@ -1,46 +1,0 @@
----
-title: Корупцијата и подмитеноста е израз на интелектуална немоќ
-cover: '/images/uploads/article-4.jpg'
-date: 2020-01-26
-author: ['Maja Terziova', 'Adem Jashari']
-published:
- Diversity Media: 'http://diversitymedia.mk/%d0%ba%d0%be%d1%80%d1%83%d0%bf%d1%86%d0%b8%d1%98%d0%b0%d1%82%d0%b0-%d0%b8-%d0%bf%d0%be%d0%b4%d0%bc%d0%b8%d1%82%d0%b5%d0%bd%d0%be%d1%81%d1%82%d0%b0-%d0%b5-%d0%b8%d0%b7%d1%80%d0%b0%d0%b7-%d0%bd%d0%b0/'
----
-
-Пишуваат: Маја Терзиова и Адем Јашари
-
-„Личноста, која конфликтните состојби сама си ги креира и режира, се нејзина насушна и душевна храна со заробен ум и најчесто страдаат многу личности, кои се наоѓаат околу нив. Никогаш не покажува желба и склоност за разбирање во интерперсоналните односи,со меѓусебна почит и внимание, никогаш нема чувство за вина,туку обратно – во секоја конфликтна ситуација умее да ја обвини секогаш другата страна и многу свирепо се одмаздува“.
-
-Методи Чепреганов е специјалист по невропсихијатрија и епилептологија со супспецијализација на компјутерска томографија и магнетна резонанца и има заедничка медицинска литература со угледни француски професори. Повеќе од 25 години учествувал во наставата како професор во областа на неврологија и психијатрија.
-Своите тези ги објаснува на многу едноставен начин, без стручни термини и нејасни зборови за да и ги долови на широката публика. Освен што е многу познат стручњак и по особено актуелни и атрактивни теми за денешно време, иако е во деветтата деценија во животот, Чепреганов се’ уште е активен лекар, држи популарни предавања за стресот и другите болести на денешното време. Неговите предавања за стресот и депресијата имаат многу голема посетеност.
-Со професор Чепреганов разговаравме за тоа колку и како стресот влијае на нашето здравје, како најчесто организмот реагира при стресни состојби,кои се последиците и како може да се справиме во стресни ситуации. Чепрганов ни објасни и за психичките пречки, агресивноста, гордоста, зависта и нарцисоидноста.
-
-Како би го дефинирале деструктивниот нагон, а како стресот како проблем со кој се соочува голема процентуалност на населението?
-
--Деструктивниот нагон е ендоген, генетски наследен фактор, кој со добра контрола можеме да го доведеме до рационална стабилност, додека стресот е психо трауматски фактор, кој го напаѓа интегритетот на личноста и ги пореметува параметрите на нормалнотото постоење и доведува до психосоматски болести.
-
-Што означува акумулација на стресоген фактор и дали има решение и одговор?
-
--Акумулација на стресогени фактори се егзогени психо трауми,кои прават револуција во биолошкиот метаболизам во нашиот организам и за жал секоја пролонгирана стресогена состојба неминовно води до генеза на психосоматски болести.Одговорот лежи во нас, во нашите карактерни особини и во нашата способност да го контролираме стресот и деструкцијата што ја носиме во нас, во името на страста на алчност за пари, моќ и за слава.
-
-Што вели невронауката во однос на агресијата, ароганцијата и насилството?
-
--Тоа се случува доколку личноста има карактерни особини, без капацитет на соодветна контрола на емоциите, со интелектуална немоќ,со незрелост, со незнаење и привитизам.На пример и корупцијата и подмитеноста претставуваат интелектуална немоќ на заробен и профитерски ум со недостаток на морални, социјални, духовни и етички норми.Агресијата како интелектуална немоќ за справување со соговорникот е отсуство на знаење,немање ментален капацитет и концентрација, култура на комуницирање и однос кон дијалогот во меѓучовечките односи.
-Во нашиот мозок постојат две спротивставени структури.Едната е нашата емоционална интелеггенција,нашите емоции и емотивен живот кај секого од нас, додека другата ги контролира нашите емоции, ја носи нашата одлука во мисловен процес, со каков вокабулар, но и како и на кој начин, со каква мимична гестикулација,гримаса, боја на глас и движење ќе биде нашиот однос наспроти соговорникот, дома на улица, на работа и сл.
-
-Кој профил на личност го сметате за најпроблематичен?
-
--Нема полоша работа од самољубието, гордоста, зависта или, со еден збор, од нарцисоидноста и паранојата како погрешна мисла.Нарцисоидноста означува човечки деструктивен нагон, хипертрофирано его (со суета, нетолерантност и инает), извршување на задоволство од суета, егоистичен восхит на сопствените физички или ментални атрибути кои се манифестираат со арогантност и гордост.Таквата личност живее за конфликти и се храни од нив.Конфликтните состојби кои сама си ги креира и режира, се нејзина насушна и душевна храна со заробен ум и најчесто страдаат многу личности кои се наоѓаат околу нив.Никогаш не покажуваат желба и склоност за разбирање во интерперсоналните односи, со меѓусебна почит и внимание,никогаш немаат чувство за вина, туку обратно, во секоја конфликтна ситуација умеат да ја обвинат секогаш другата страна и многу свирепо се одмаздуваат.
-Сепак, многу по опасна е групната нарцисоидност, која создава еден многу опасен феномен на солидарност и кохезија на една одредена популација. Созреаната група е многу лесна за манипулација и злоупотреба, ништо не може да ја стопира, особено кога ке се појави градација, фанатичност и фанатизам. Разумот едноставно исчезнува. Има многу примери во светската историја, кои на дело ги покажуваат дејствијата на масовната хистерија, како војни, штрајкови, протести, навивање со дивеење по стадионите и сл.
-
-Од што зависи психо профилот на една нација ?
-
--Од сите архетипови на колективното јас, од што покасно ке зависат менталните и интелектуалните склопови и капацитети на општиот позитивен, или негативен квантум на таа нација, популација или народ, неговото однесување, интелектуален капацитет и знаења, работоспособност, толерантност и флексибилност, визионерство и креативност, со еден збор – нивото на неговото постоење, крајно примитивно од една и високо интелектуално, практично, хумано и издржано од друга страна.
-
-Има ли решение за човековата деструкција?
-
-Теориски да, но, таа е можна само во корелација со хармоничноста и добро осмислена и програмирана висока свесност на еден психо профил, во зрело општество со високи воспитни, едукативни, социјални, правни, економски, морални, етички и духовни норми со богат резервоар на знаења.
-
-Од што зависи превенцијата на агресијата ?
-
-Најпрво од едукацијата и воспитанието, семејството, работното место, условите на живот, ангажираноста, мотивацијата за работа и спортувањето како многу важен фактор.
