--- content/portfolio_mk/kur-je-ne-karroce-gjerat-qe-ne-maqedoni-jane-luks-ne-gjermani-jane-normale.md
@@ -1,54 +1,0 @@
----
-title: Кога сте во количка „работите во Македонија се луксуз, а во Германија се нормални“
-cover: '/images/uploads/kur-je-ne-karroce-gjerat-qe-ne-maqedoni.jpg'
-date: 2020-02-26
-author: ['Rita Behadini', 'Dajana Lazarevska']
-published:
- Diversity Media: 'http://diversitymedia.mk/%d0%ba%d0%be%d0%b3%d0%b0-%d1%81%d1%82%d0%b5-%d0%b2%d0%be-%d0%ba%d0%be%d0%bb%d0%b8%d1%87%d0%ba%d0%b0-%d1%80%d0%b0%d0%b1%d0%be%d1%82%d0%b8%d1%82%d0%b5-%d0%b2%d0%be-%d0%bc%d0%b0%d0%ba%d0%b5/'
-
----
-
-Пишуваат: Рита Бехадини и Дајана Лазаревска
-
-Пред некое време видов видео на некои активисти во Тетово, кои покажаа колку е тешко да се движиш со количка. Ова ме натера да размислам за сите тешкотии, со кои овие луѓе се соочуваат во секојдневието, а од друга страна, со само малку грижа, можат многу да му понудат на општеството.
-
-Еден од нив е и Гадаф Мемеди. Гадаф е првиот студент со посебни потреби од Македонија, кој ја искористил можноста на Еразмус програмата за размена на студенти во странство. Тој е студент на Универзитетот во Тетово, на студиската програма за германски јазик и одлучил размената да ја оствари во германскиот град Марбург. Со инвалидската количка со која се движи, во Македонија најчесто, Гадаф има потешкотии со пристапот до училишните објекти, институции или тротоарите. Зградите, улиците, превозните средства немаат рампи за лицата со посебни потреби. А според Гадаф, додека ова се смета за луксуз во Македонија, во Германија е нормално, посебно транспортните рампи и чистиот воздух. Тој вели дека студентите треба да ги искористат студиските програми за размена и дека ако аплицираат за истото, нема ништо да загубат, туку напротив само ќе добијат.
-
--Од кои причини аплицираше за студентската размена во странство?
-
-„Подолго време кај мене е присутна идејата за волонтирање и вклучување во нови активности. Ја практикував преку размените и обуките на организацијата “ЛОЈА“. На некој начин, преку странските доброволци кои доаѓаа во оваа организација, светот дојде во Тетово. Единствената разлика овој пат беше што ние отидовме во светот“.
--Како ти се променил животот откога замина да студираш во странство?
-
-„Многу работи се променија. Запознав луѓе од целиот свет. И што повеќе луѓе запознаваме, сфаќаме колку е мал светов“.
-
--Дали забележуваш разлики во образовниот систем таму и колку ти беше тешко да се прилагодиш?
-
-„Има многу разлика во целиот образовен процес тука, но најважното кое би го спомнал е фактот дека има многу повеќе посветеност. Постои чувството дека се’ е на своето место“.
-
--Кои се најголемите разлики меѓу Германија и Македонија во услугите и системот воопшто?
-
-„Се’ е коректно и функционално овде (Германија н.з.), како што и треба да биде. Да имаш работи кои функционираат како што треба во Македонија, се смета за луксуз. Овде е едноставно уште една во низата на нормалните работи. Разликата е што кај нас уште предмет на пазар дали ќе се уживаш едно човеково право или не. Овде не е во прашање дали ќе имаме патишта или не, дали ќе имаме чист воздух или не. Тука е во прашање во која насока ќе се развива земјата“.
-
--Дали ти се олесни животот таму, од аспект на пристап до јавни услуги и институции?
-
-„Да, ми се олесни, бидејќи како што кажав претходно, сѐ е на своето место. Рампите во јавниот превоз и чистиот воздух не се луксуз, за кој треба да се преговара. Тоа е право кое треба му се обезбеди на секој човек“.
-
--Какво е чувството на искусување на нови култури?
-
-„Со текот на времето кое го поминувам овде, искусив како е да се сретнеш со нови култури, како е да бидеш вклучен во работи, за кои многу луѓе мислат дека не можат да ги направат и разбрав дека јазикот и културата се дел од нашата суштина. Но, во основа сите сме исти. Имаме различни јазици, но нашите чувства се исти. Можеби зборуваме различни јазици, но плачот и смеата се исти насекаде во светот. Можеби културата нè разликува, но во суштина сме исти, сите сме луѓе“.
-
--Според тебе, зошто студентите треба да ги искористат можностите за студентски размени?
-
-„Затоа што тоа е нешто што секој студент треба да го искуси, бидејќи е нешто кое ги менува луѓето на подобро. Тоа е искуство кое ти го менува животот, без разлика каде одиш и што правиш. Ти се дава можност да запознаеш еден поинаков дел од светот. Не губиме ништо од себе ако одиме на студентски размени,можеме само да добиеме нешто дополнително“.
-
--Сега, откога си живеел и во Германија, дали пак би се вратил да живееш во Македонија?
-
-„Ќе мора да се вратам, бидејќи така пишува во мојот пасош. Но, ова е правен апсект. Како прво, тоа е земјата каде што сум роден, а како второ, тоа е земјата каде што сум доживеал многу искуства. Кога би требало да изберам, би ги избрал двете. Би сакал да донесам дел од Германија таму, но исто би сакал и ним да им дадам дел од нас“.
-
--Што е она што многу ти се допаднало таму и би сакал да биде исто и во Македонија?
-
-„Свеста на луѓето, фактот дека секој ја сака својата земја и секој го има чувството дека треба да следи некое правило едноставно бидејќи постои, а не затоа што некој ќе го казни. Тука сум видел луѓе кои чекаат на семафор во три часот по полноќ, иако не поминува ниту едно возило“.
-
--Што би им порачал на студентите?
-
-„Направете сѐ што мислите дека не може да го направите. Едноставно, направете го! Не треба прашањето да биде дали ќе можам, туку како? И, ако стигнете до точка кога мислите дека не можете сами да постигнете нешто, тогаш запомнете дека не сте сами и дека имате луѓе околу вас. Проблемот може да биде еден, но луѓето околу вас се милиони и имаат методи за решавање на еден проблем, значи дека се во многу поголем број отколку проблемите. Значи, продолжете напред, размислете што може да направите сами, и потоа ќе разберете дека не сте сами“.
