--- content/portfolio_mk/me-nenvleresim-te-bulizmit-deri-ne-degradim-te-karakterit-te-femijeve.md
@@ -1,44 +1,0 @@
----
-title: Со потценување на булизмот до деградирање на детскиот карактер
-cover: '/images/uploads/bullying.jpg'
-date: 2020-01-26
-author: ['Dorina Ajdini', 'Maja Terziova']
-published:
- Diversity Media: 'http://diversitymedia.mk/%d1%81%d0%be-%d0%bf%d0%be%d1%82%d1%86%d0%b5%d0%bd%d1%83%d0%b2%d0%b0%d1%9a%d0%b5-%d0%bd%d0%b0-%d0%b1%d1%83%d0%bb%d0%b8%d0%b7%d0%bc%d0%be%d1%82-%d0%b4%d0%be-%d0%b4%d0%b5%d0%b3%d1%80%d0%b0%d0%b4%d0%b8/'
-
----
-
-Пишуваат: Дора Ајдини и Маја Терзиова
-
-Запрепастувачки голем е процентот од средношколците во Тетово кои биле жртви на булинг во училиштата од нивни врсници, од професори, но и од највисоките органи во образованието. Ова го покажува истражувањето на претседателот на Унијата на средношколци (УСТ) во Тетово, Бленди Ходаи и неговите колеги. Според него, булингот е тема за која треба почесто да се зборува, затоа што не можеме да ја надминеме ако премолчуваме. Но, наместо поддршка, Ходаи бил непријатно изненаден од односот на директорите на училиштата, од кои добил закана по објавувањето на резултатите од истражувањето.
-
-![Bullying](/images/uploads/bul-1.jpg)
-
-Спроведувачите на истражувањето направено меѓу 1729 ученици на возраст од 15-19 години, доживеале уште едно изненадување. Нивните очекувања биле дека физичкиот булинг ќе биде позастапен кај учениците. Но, резултатите од ова истражување покажаа дека 79 отсто од испитаниците сметаат дека таканаречениот невидлив булинг или психички булинг е повеќе застапен во односот наставник – ученик, но и кај самите ученици.
-
-„Многу загрижувачки податок што произлезе од истражувањето, е дека 44 отсто од учениците биле жртви на булинг. Тој број веројатно е поголем, бидејќи булизмот станува секојдневна рутина во нашите училишта“, рече Бленди Ходаи , лидер на УСТ.
-
-
-**Што е булинг и како да го препознаеме?**
-
-Булинг претставува навредливо однесување кон одредени лица, кое се повторува, е штетно и агресивно. Варира во однос на нивото на насилството и можате да биде и физички и психички. Темата за булингот е тема за која многу малку се дебатира кај нас. На булингот не му се дава голема важност и се занемарува, не само од родителите и наставниците, туку и од државните институции.
-Но, како да знаеме дека едно дете е жртва на булингот? Најчестите знаци се кога одбива да оди на училиште, со изговорот: „Не сакам да одам на училиште, ме боли стомакот, не се чувствувам добро!“. Психолозите предупредуваат ‒ ако откриеме дека детето дава вакви знаци и тие се чести, случајот треба да се испита.
-
-![Bullying](/images/uploads/bul-2.jpg)
-
-Фиторе Бајрами, професорка по психологија на Тетовскиот универзитет, објаснува дека булингот се дефинира како намерно агресивно однесување.
-„ Лицето кое има намера да задева друго лице, тоа го прави со целосна свест. Сака да му наштети на другиот, почнувајќи од тоа „да прави да се чуствува лошо“, продолжувајќи во други форми. Тоа што е карактеристично кај булингот е што лицето кое предизвикува булинг, во овој случај се нарекува агресор и секогаш има поголема сила од другиот“, вели професорката Бајрами.
-Посебен фокус треба да се посвети и на изворот на овој феномен, што се јавува во нашето општество. Психологот Едлира Дестани вели дека внимателно треба да се следи секоја промена кај учениците и да се најде изворот на тоа однесување.
-„Во одредени моменти кога не сме во состојба да се изразиме, применуваме такви механизми што ги задеваме другите за да се воздигнеме себеси. Ова се случува со сите генерации. Во меѓувреме, за да се издигнеме себеси, започнуваме со вербално вознемирување, вербално исмејување, со што другото лице ,кое е жртва во овој случај, не се спротивставува на задевањето и избегнувајќи го, преминува во фазата во која почнува да се повлекува. За да го избегнеме булингот, ние треба секое дете на многу мала возраст да го оставаме да се искаже за сѐ што го мачи. Секогаш треба да се земе сериозно случајот и да се најде изворот зошто тоа лице се однесува така,“ вели Дестани.
-За да се реши проблемот помеѓу врсниците на училиште, смета психологот Дестани, децата не треба да се срамат да проговорат за инцидентот со своите родители, да добијат поголема поддршка од нив и поголема ангажираност од психолозите. Поголем акцент и редовна грижа за менталното здравје. Треба секој да се стреми кон тоа да има поголема самодоверба, да не дозволи некој друг потценувајќи го да се воздигне себеси. Наместо да критикуваме, да охрабруваме.
-Но, колку се професорите се запознаени со оваа појава и како реагираат во случај да видат ученик кој задаева или се изживува со некој соученик?
-УСТ вперува прст во психолозите во училиштата кои, според нив, треба да сносат одговорност за оваа состојба, бидејќи од нивна страна никакво истражување не е направено. Од Унијата велат дека психолозите дури не се ни запознаени со терминот „булинг“. Тие ова го нарекуваат дисфункција на канцеларијата на училишните психолози. Најголема иронија, според УТС, е што учениците не ни знаат кои се психолозите во нивните училишта.
-
-![Bullying](/images/uploads/bul3.jpg)
-
-**Булингот ‒ сѐ уште табу тема во општеството**
-
-Дека за булингот не се зборува меѓу професорите во училиштата потврдува и средношколската професорка, Линдита Емерлаху, која укажува на неколку слабости во односот професор – ученик.
-„Ние, наставниците, го сфаќаме овој концепт како многу важен, почнувајќи од самите себе. Нашата главна задача како професори треба да биде фокусирање на тоа како да го зацврстиме карактерот на учениците и тие да знаат како да реагираат во случај на булинг. На многу ученици им недостасува самодоверба. Не ми се допаѓа кога учениците секогаш велат „ДА“ за сѐ што јас ќе им кажам. Треба да постои поблиска релација помеѓу учениците и професорите. Треба да постои човечка блискост, која ќе помогне во отстранување на социјалниот булинг,е ставот на Емерлаху.
-Професорката Емерлаху, смета дека против булизмот може да се бориме и со менување на начинот на предавање со примена на интерактивно предавање. На овој начин учениците ќе се вклучат во предавањето и ќе почнат да размислуваат дека нивното мислење вреди.
-Сето ова е и дел од образовната дејност, какви деца ќе воспитуваме во иднина и во какви личности тие ќе пораснат.
