--- content/portfolio_mk/miroslav-draganov-si-te-vazhdojni-me-buzeqeshje.md
@@ -1,48 +1,0 @@
----
-title: 'Мирослав Драганов: Како да продолжите со насмевка'
-cover: '/images/uploads/miroslav-draganov.jpg'
-date: 2020-01-26
-author: ['Ljubica Ivanova', 'Arta Qerimi']
-published:
- Diversity Media: 'http://diversitymedia.mk/%d0%bc%d0%b8%d1%80%d0%be%d1%81%d0%bb%d0%b0%d0%b2-%d0%b4%d1%80%d0%b0%d0%b3%d0%b0%d0%bd%d0%be%d0%b2-%d0%ba%d0%b0%d0%ba%d0%be-%d0%b4%d0%b0-%d0%bf%d1%80%d0%be%d0%b4%d0%be%d0%bb%d0%b6%d0%b8%d1%82%d0%b5/'
----
-
-Пишуваат: Љубица Иванова и Арта Ќерими
-
-Споделува лични искуства за да ги охрабри другите луѓе и да им пренесе дека е подобро нешто да се преземе како превенција и да се работи на личен развој кога има здрава основа – Мирослав Драганов, мотивациски говорник и тренер за самодоверба, поранешен извршен директор на Институтот за човекови права, кој уште на 21 година го освоил светот со натпревари на јавно говорење на исто рамниште со светски лидери. Магистрирал уставно право на Правниот Факултет во Скопје и автор е на книгите „На своја кожа – продолжи со насмевка“ и „Продолжи со насмевка – за љубов создавајќи“, кои се автобиографски и мотивациски.
-
-Покрај издадените книги, Драганов, исто така, споделува искуства и објавува видеа на социјалните мрежи, држи и работилници за личен развој и самодоверба, работи со поединци, компании, тимови. Во подготовка е и аудио верзија на неговата книга, која нема да биде во комерцијални цели, а е наменета за слепи лица.
-
-„Надворешниот свет е едно, но, како ние реагираме на надворешниот свет е сосема друго, најмалку што сакам е некој друг да го поминува тоа што сум го поминал јас“, вели Мирослав Драганов.
-Во својата прва книга раскажува лични и трагични моменти, кога, како што вели, не му бил потребен само мотиватор туку и бил и во ситуација да дојде до психијатар. Тој вели дека животот буквално го трескал од земја.
-Со него разговаравме за тие пресвртни, клучни моменти.
-
-Што е тоа што вас ви дава сила да продолжите понатаму ако ви било тешко?
-
--Животот никогаш нема да ви биде совршен, но, важен е фокусот, а вие бирате на што ќе се фокусирате, вели Мирослав и се присетува на ситуација како само затоа што верувал на даден збор, бил донесен во тешка животна ситуација. Сепак, ситуација од која доста научил.
--Можеби сум еден од ретките кој во пет години дал четири откази, во време во кое велат дека тешко е да се најде работа – вели тој и раскажува како откако му било ветено работно место, а поради што напуштил сигурна работа, работодавачот не го одржал ветениот збор, а Мирослав останал без работа.
-
-„Мојата животна девиза е дека за се постои решение, освен мртов човек да оживее“.
--Тоа беше тежок период затоа што ниту татко ми ниту маќеа ми не работеа, решив да не им кажувам ништо, тие мислеа дека две недели јас одам на работа, излегував дотеран од дома, знаев дека не може да ми помогнат, а само ќе се нервираат, а знаев и дека ќе најдам решение, па дури тогаш ќе ги известам. Од денешна перспектива, мило ми е што и тоа искуство сум го поминал, бидејќи тој ги затвори канцелариите, а јас сум таму кај што сум и допрва растам, објаснува Драганов.
-
-Од правник, до мотивациски говорник. Наоѓате ли поврзаност во овие две професии, бидејќи не секој може да биде мотивациски говорник? Кога вие првпат почувствувавте дека тоа може да биде ваша професија?
-
--Кога можат тие зошто не и јас. Се фокусирав целосно на работата како тренер за самодоверба и мотивациски говорник. Сфатив дека ова ме прави посреќен, колку и да изгледа несигурно во очите на останатите. Официјално го работам тоа што го сакам, иако сум го правел низ годините, без да наплаќам, приватно со мои блиски луѓе, кои отсекогаш ми велеле дека им влијаам позитивно и ме поттикнувале. Правото не е целосно занемарено во сегашната моја работа, бидејќи ова е превенција, да си ги изградите односите со луѓето, со вработените, да се едуцирате, да знаете кои се вашите човекови права да не дозволите некој да ви ги наруши. Меѓудругото, позитивна едукација во еден тим води до поголема продуктивност, негативната води кон мобинг, што е нарушување на човекови права, смета Мирослав Драганов.
-
-Драганов вели дека следел голем број странски мотивациски говори, но преживеал и доста предизвици во животот. Смета дека наученото може да го применува, додавајќи елемент на дополнителна вредност така што комбинира теорија и пракса. За себе вели дека ги поминал сите скалила, почнал од најниско – приправник, па се до највисоко, до извршен директор.
-Реакциите од поединци, вели дека најмногу го радуваат. Го истакнува примерот со магионичарот, Кристијан Шопов кој претежно работел со деца, но благодарејќи на Мирослав, ја остварил животната страст да работи со возрасни, пребродувајќи го стравот од јавни настапи пред голема публика, а Драганов е најблагодарен кога таквите искуства ги споделуваат и ги пренесуваат на други луѓе.
-
-Може ли секој човек да има самодоверба? Како вие низ годините сте ја изградиле вашата самодоверба, што за вас претставува самодовербата?
-
--Верувајте во себеси и во вашите способности, но не се преценувајте.
-Постојат две крајности. Едниот момент е да се биде самобендисан и нарцисоиден, а другиот да не веруваш во себеси со што нема да ги осознаеш твоите способности. Самодовербата е на средината и е најубава, и додава дека таа е пропорционална со тремата, така што кога имате повеќе трема, ќе имате помалку самодоверба, советува Драганов.
-
-Тој раскажува за негова колешка која ги имала изгледано сите ТЕД ТАЛКС, меѓутоа кога требало таа да излезе да зборува на конференција, имала трема и не било успешно, затоа Драганов вели дека една од техниките за градење самодоверба е со вежбање.
--Ако имате трема од нешто, правете го повеќе пати. Учете како да го подобрите, но не само со теорија, советува Драганов. И заборавите на вашето его, додава тој.
-
-Драганов ни откри и дека често му велат дека не размислува како луѓето во Македонија, па го прашавме колку кај нас има фидбек на тоа што го прави.
--Во Македонија не секогаш ни недостасува квалитет, туку да бидеме поотворени за промените. Луѓето кај нас некогаш сакаат прво да видат дали има резултати, вели тој и додава дека во поново време и тие што биле скептични, кога ќе ги почувствуваат резултатите, препорачуваат.
-
-Совет, како граѓаните да продолжат со насмевка, во услови на суров свет, брз живот и проблеми?
-
--Времето си е време, не можам да го сменам, но можам да го сменам моето однесување. За почеток нека сменат приоритети. Ајнштајн вели ако со еден начин на размислување создавате проблем, не очекувајте да го решите со истиот начин на размислување. Самиот начин на размислување нема да ви донесе многу ако не е проследен со акција. Јас сум тој кој вели обидувајте се, грешките се искуство, ама ако сакате да научите, да бидете отворени за знаење да се надградувате во вештини. И работете го она што ве прави среќни, не само на работа туку цело време. И продолжете со насмевка, е пораката на Мирослав Драганов, мотивациски говорник и правник.
