--- content/portfolio_mk/perfaqesimi-gjinor-ne-vend-nen-standardet-nderkombetare.md
@@ -1,25 +1,0 @@
----
-title: Родовата застапеност во земјата под меѓународните стандарди
-cover: '/images/uploads/women-rights.jpg'
-date: 2020-01-26
-author: ['Teuta Buchi', 'Renata Penchova']
-published:
- Diversity Media: 'http://diversitymedia.mk/%d1%80%d0%be%d0%b4%d0%be%d0%b2%d0%b0%d1%82%d0%b0-%d0%b7%d0%b0%d1%81%d1%82%d0%b0%d0%bf%d0%b5%d0%bd%d0%be%d1%81%d1%82-%d0%b2%d0%be-%d0%b7%d0%b5%d0%bc%d1%98%d0%b0%d1%82%d0%b0-%d0%bf%d0%be%d0%b4-%d0%bc/'
-
----
-
-Пишуваат: Теута Бучи и Рената Пенчова
-
-Родовата еднаквост и соодветната родова застапеност во јавните институции изгледа остануваат не толку блиска реалност на владата, која тврди дека ги вклучила сите мотори за усогласување на стандардите и вредностите со Европската унија, за да потоа стане и нејзина членка. Од она што успеавме да го анализираме, преземањето на владиното кормило од една левичарска партија со социјалдемократска ориентација, не се одрази во намалување на родовите разлики. Што се однесува до ова прашање, бидејќи досега ништо конкретно не е направено во оваа насока.
-
-На чело на јавните претпријатија продолжува назначувањето на мажи и раководните позиции се понедостапни за жените. Според податоците добиени од Народниот правобранител, повеќето јавни претпријатија немаат соодветна родова застапеност. Во трите претпријатија кои ги земавме за анализа, се забележува дека на именувануваните функции и раководните позиции главно доминираат мажите. Во Македонската радио и телевизија (МРТВ), двете именувани позиции ги држат мажи. Од 117 раководни позиции, 65 се жени. Од вкупниот број на 794 вработени, 55,29% се мажи, додека 44,71% се жени.
-Во Јавното претпријатие за државни патишта, за раководител е назначено лице од машки пол, а од 92 раководни позиции, 74 или 80.43% се мажи, додека 18 или 19,57% се жени. Вкупниот број на вработени во ова претпријатие е 339, од кои 80,83% се машки лица. Исто така, и Јавното претпријатие Македонски железници – Инфраструктура, нема соодветна родова застапеност. Со претпријатието раководи лице од машки пол. На раководни позиции, од 62 работни места, на 72,58% се вработени машки лица. Што се однесува до вкупниот број на вработени, од 984, 909 или 92,38% се мажи.
-Од ЈП за државни патишта, објаснувајќи ја причината зошто е ваква состојбата на родовата застапеност, велат дека на нивните конкурси, главно, се пријавуваат машки лица, додека, според нив, се трудат да бидат повнимателни за раководните позиции, каде вработуваат и жени.
-
-„Во Јавното претпријатие за државни патишта најголем број вработени ги извршуваат должностите на наплатни станици на територијата на Република Северна Македонија. Со оглед на условите за работа (работа во смени, ноќни смени …) машките се позаинтересирани за овие работни места. Во управата на Јавното претпријатие за државни патишта, во однос на стручните служби, голем број на високи раководни позиции се пополнети од женски лица“, вели Билјана Цветковиќ, раководител на секторот за човечки ресурси и општи работи на ова претпријатие.
-Хелсиншки комитет за човекови права, оценува дека застапеноста на жените на раковдни функции во јавниот, но и приватниот сектор, е многу мала и е доста незадоволителна. „Имајќи во предвид дека само 15% од министерските позиции им припаѓаат на жените и дека од 80 градоначалници, само 6 се жени, доаѓаме до заклучок дека учеството на жените во политичкиот живот е на многу ниско ниво. Иако немаме точни податоци какво е застапувањето на жените во државните институции, искуствата кажуваат дека, и на јавниот и на приватниот сектор, родовата еднаквост и рамноправната застапеност на жените не е реализирано. Таму каде што има приближување на рамноправна родова еднаквост, се забележува дека на повисоките работни позиции се појавува голема неусогласеност во корист на машките, додека застапеноста на жените на административни позиции е повисока“, велат од Хелсиншкиот Комитет.
-
-Иако бројот на женските лица кои дипломирале на различни факултети е висок, бројките значително се намалуваат кога станува збор за материјализирање на дипломите во работни места. Женските кадри од различни профили исчезнуваат на пазарот на трудот, доминиран од машките.
-
-„Иако податоците покажуваат дека поголем процент на жени (59%) отколку мажи (41%), завршуваат високо образование, жените подоцна исчезнуваат од пазарот на трудот, односно вработувањето на жените (34,6%) се намалува во однос на вработувањето на машките (53,6%). Ова укажува на фактот дека не станува збор за недостаток на образувани жени и кадри од женскиот род, туку за постоење на бариери при пристап до пазарот на трудот како и понатамошното напредување во кариерата. Ова е како последица на традиционалните општествени родови норми, кои сеуште имаат влијание врз учеството на жените на пазарот на трудот, имајќи го предвид балансот кој жената треба да го направи меѓу приватниот и професионалниот живот“, оценуваат од Хелсиншкиот Комитет.
-„Еднакви права за мажите и жените“ е едно од основните начела на Повелбата на Обединетите нации, усвоена од светските лидери во 1945 година. Принципот на родова еднаквост признат и во Универзалната декларација за човекови права во 1948 година, која ги потврди „еднаквите права на мажите и жените“. Интегрирањето на родовата перспектива е процес на проценка на импликациите за жените и мажите за какво било планирано дејство, вклучително законодавство, политики или програми, во секое поле и на сите нивоа.
