--- content/portfolio_mk/shtohet-denoncimi-i-rasteve-te-dhunes-ne-familje.md
@@ -1,39 +1,0 @@
----
-title: Се зголемува бројот на пријавени случаи на семејно насилство
-cover: '/images/uploads/shtohet-denoncimi-i-rasteve-ne-familje.jpeg'
-date: 2020-03-26
-author: ['Teuta Buchi', 'Evi Shkopi']
-published:
- Diversity Media: 'http://diversitymedia.mk/%d1%81%d0%b5-%d0%b7%d0%b3%d0%be%d0%bb%d0%b5%d0%bc%d1%83%d0%b2%d0%b0-%d0%b1%d1%80%d0%be%d1%98%d0%be%d1%82-%d0%bd%d0%b0-%d0%bf%d1%80%d0%b8%d1%98%d0%b0%d0%b2%d0%b5%d0%bd%d0%b8-%d1%81%d0%bb%d1%83%d1%87/'
-
----
-
-Пишуваат: Еви Шкопи, Теута Бучи
-
-Последниве години процентот на пријавени случаи на семејно насилство е зголемен. Ана Аврамоска – Наушкова има прилика директно да се справи со такви случаи, бидејќи е правен советник во Националната мрежа против насилство врз жените и семејно насилство. Неретко овие случаи имаат емотивно влијание врз неа. Аврамоска – Наушкова вели дека жените се најчести жртви на насилство. Ситуацијата е дури посложена кога се финансиски зависни од насилникот, па ова дополнително го отежнува нивното напуштање на насилната средина. Во случаите на семејно насилство, посебно децата имаат отежната психолошка состојба.
-
-Следува целосното интервју со Аврамоска – Наушкова.
-
-Дали имате бројки кои кажуваат каква е состојбата во Македонија што се однесува до семејното насилство?
-
--Имаме неколку бројки, кои се обезбедени од некои службени институции како Министерството за внатрешни работи. Во годишниот извештај за 2018 година се вели дека 1006 случаи на семејно насилство се пријавени. Од Министерството за труд и социјална политика, за 2018 год. се пријавени 919 нови регистрирани случаи. Ако се споредат извештаите на Министерството за внатрешни работи за 2017 и 2018, се гледа дека пријавувањето на семејно насилство се зголемило за 11%.
-
-Кој е најчесто жртва на семејно насилство?
-
--Најголем дел од овие пријави на семејно насилство е врз жени, кои се најчесто жртви на семејно насилство и тоа во 85% од случаите се жртви на физичко насилство, од нивниот интимен или брачен партнер. Од податоците на Министерството за труд и социјална политика за 2018, најголемиот дел на случаи се пријавени од жени, а потоа од мажи и деца. Имаме една мала статистика, која успеавме да ја обезбедиме во септември, преку правото за слободен пристап до информации од јавен карактер. Значи, побаравме податоци од МТСП за статистика за семејно насилство во период јануари-јуни 2019 година. Бројката е 767. Дали сега е зголемена храброста за пријавување на такви случаи, или е зголемено семејното насилство, можеби некое друго истражување може да даде одговор на ова.
-
-Што конкретно вие преземате како организација кога една жена, жртва на насилство ви се обраќа за помош?
-
--Ќе ви кажам за конкретен случај. Една жена, мајка на две малолетни деца, жртва на семејно насилство, се обрати кај нас. По дискусијата со правниот советник, се дојде до заклучок дека може да се преземат повеќе правни постапки за финансиско зајакнување на жената, во согласност со Законот за социјална заштита и Законот за заштита на деца. Исто така, во тек на тој период, жената која беше жртва на семејно насилство, имаше итна потреба за храна и хигиенски средста, па и се обезбеди пакет со основни средства, преку Интервентниот женски фонд. Во соработка со други организации, госпоѓата доби и облека и за себе и за своите две малолетни деца.
-
-Дали жените се плашат да го пријават насилството кое се врши врз нив?
-
-– Според последното истражување на ОБСЕ за 2018 година, во кое беше опфатен целиот Западен Балкан, вклучително и Македонија, се покажа дека 42% на жени кои се жртви на семејно насилство, поради фактот дека немаат финансиски можности, немаат имот или немаат поддршка од блиските, или не ја напуштаат средина каде се подложени на насилство, или ја напуштаат па пак се враќаат. Исто така, врз колебањето за пријавување на насилство влијаат и предрасудите на општеството, озборувањето на луѓето, покажувањето со прст, а во помалите средини ова е уште поизразено бидејќи сите се познаваат меѓусебно. Загрижувачко е високото ниво на недоверба во институциите, што не секогаш е оправдано. Мислам дека главниот проблем на жртвите на семејно насилство е, како прво, нивната психолошка состојба, потоа нивната финансиска состојба и правната помош која треба да им се дава. Ако самите отворат постапка, за ова е сигурно дека ќе има материјални трошоци. Меѓутоа, доколку Центарот за социјални работи потврди случај на семејно насилство, по службена должност, треба да започне процедура за изрекување на потребни мерки.
-
-Според информациите кои ги имате, дали надлежните институции постапуваат навремено кога се пријавува случај на семејно насилство?
-
--Групите каде сме фокусирани, во согласност со спроведената анализа со жени жртви на семејно насилство, изразија поплаки или одредена незаинтересираност на Центрите за социјална работа. Во принцип, кога имаме одбранбен механизам кој треба да се стави во располагање на жртвите на семејно насилство, главната улога го има Центарот за социјални работи, потоа полицијата и здравствените работници.
-
-Дали имате информации дали во Македонија се изградени потребните капацитети за да се заштитат лицата кои се жртви на домашно насилство?
-
-– Министерството за труд и социјална политика јавно посочи колку капацитети има вкупно, во согласност со Истанбулската конвенција, која кај нас започна да се спроведува во рамките на акциониот план од октомври 2018 година. Имаме многу ограничени капацитети. Засолништа има во 5 градови низ Македонија со капацитет од вкупно 35 легла. Ова е на национално ниво. Сметаме дека не е доволно и дека треба да се постапи во согласност со Истанбулската конвенција, да се направат прилагодувања и да се исполнат стандардите.
