--- content/portfolio_mk/ushqimi-organik-ne-vend-akoma-ne-fillesen-e-tij.md
@@ -1,48 +1,0 @@
----
-title: Органската храна во земјава – се’ уште во зачеток
-cover: '/images/uploads/organic-food.jpg'
-date: 2020-01-10
-author: ['Ljubica Ivanova', 'Arta Qerimi']
-published:
- Pina: 'https://pina.mk/2959-organska-hrana-vo-zemjava-se-ushte-vo-zachetok/'
-
----
-
-**Граѓаните малку го следат светскиот тренд на употреба на органска храна. Делумно, тоа се должи на повисоките цени на органските производи, но и на фактот дека ретки се маркетите каде може да се најде органска храна. Наспроти секојдневната препорака на нутриционистите за редовно консумирање на органски производи, отворено е прашањето колку оваа храна можат да си ја дозволат граѓаните.**
-
-Љубица Иванова и Арта Ќерими
-
-Скопјани се најголеми потрошувачи на органската храна, која пак доаѓа од внатрешноста на државата, велат производителите. Еден од нив е Ангелчо Димитриев од Штип, кој заедно со неговата фамилија, десетина години се занимава со органско производство, а од неодамна и преработка на цреши, праски, кајсии, јаболка, круши, бадеми, лешници. Започнале со помали земјоделски плантажи, сега веќе работат на површина од 25 хектари во пробиштипско Зарапинци, на посед помеѓу Свети Николе и Пробиштип, во чиста незагадена средина.
-
-„Малку повеќе време е потребно за органско производство во споредба со конвенционалното. Нашите производи ги дистрибуираме во продавници за здрава, органска храна во Скопје, за цените не прават проблем, веднаш ги откупуваат органските производи. За разлика од Скопје, каде што има поголема побарувачка, во другите градови во Македонија, заинтересираноста е помала. Се уште не е развиена свеста за органска храна, луѓето се сомневаат, па затоа е потребно повеќе промовирање, но има надеж“, вели Ангелчо Димитриев – земјоделец.
-
-Семејството Димитриеви, во договор со туристичка агенција од Скопје, планира да организира посети до нивните органски плантажи, за да се развие селскиот туризам и да се промовира органската храна, за луѓето да се уверат во квалитетот на производството.
-
-Тие поседуваат и сертификат од компанијата „Балкан Биосерт“, која врши контрола на нивното производство, за што на компанијата и плаќаат по 40.000 денари од контрола, а државата на производителите им враќа 50% од сумата.
-
-За разлика од светските трендови, земјава речиси ја нема во светските статистики, кои вбројуваат држави со напредно производство и побарувачка на органска храна. Потрошувачката на органски производи не го следи трендот на државата, која го стимулира органското производство со државни субвенции, кои се повисоки за 30 отсто во однос на конвенционалното. Преку Агенцијата за финансиска поддршка во земјоделството и руралниот развој во октомври 2019 година, на сметките на 638 одгледувачи на органско производство,се исплатени над 82 милиони денари(МЗШВ).
-
-**Органска храна има само на одредени локации**
-
-Деканот на Земјоделски Факултет при УГД – Штип проф. д-р Љупчо Михајлов го следи трендот на потрошувачката на органски производи во малопродажбата.
-
-„Инцидентно, во поголемите маркети, може да се забележи постоење на одредени обележани места за продажба на етикетирани сертифицирани органски производи, но континуирана понуда и побарувачка на органска храна и сериозно развиен и организиран пазар и берза на истата за жал сеуште не може да се констатира во нашата држава“, објаснува професорот Михајлов.
-
-Повеќето големи ланци на маркети, не држат органски производи на сите свои продажни места. Врз основа на фреквентноста и побарувачката на потрошувачите се одлучува на која локација ќе се продаваат органски производи.
-
-Според една сопственичка на продавница за органска храна во центарот на Скопје, поради повисоките цени на органските производи, мала е побарувачката на растителни органски производи на домашниот пазар.
-
-„Цените за органски производи во однос на неоргански понекогаш се дури и дупло повисоки, но имаме клиенти кои користат органски производи во секојдневието, се навикнуваат на нив и ги купуваат без обѕир на цената“, вели сопственичката на органска продавница.
-
-Европскиот пазар на органски производи во 2017 годинапораснал за повеќе од 37 милијарди евра, а секој Европјанин троши по 67 евра на органска храна.
-
-**Производството се уште во зачеток**
-
-„Добра Земја“е една од продавниците кои откупуваат локална македонска органска храна.Иво Бару, вработен во продавницата, вели дека повеќето производи се откупени од производители од околни села во Кочани, Струмица, Кавадарци, Неготино, Ресен, Валандово, како и скопски села, од производители кои поседуваат сертификат за органско производство, но за да се уверат во тоа, најчесто и лично ги посетуваат производителите пред да започнат соработка со нив. Во однос на цените не постои одредена формула по која се одредуваат. Цените, според Бару, претежно зависат од тоа колку е достапен производот на пазарот, како и од цената што самите производители ја утврдиле.
-
-*„ Органското производство во Македонија како да е уште во зачеток. Во најголем број и најверни потрошувачи во нашата продавница се претежно мајки, кои купуваат органска храна за нивните мали дечиња. Покрај мајките, како консументи на органска храна се јавуваат и луѓе со здравствени проблеми, луѓе кои се придржуваат на одреден вид исхрана, водат здрав живот и кои се на диета. Продавницата постои пет години и за жал не можеме да се пофалиме со пораст на побарувачката колку што ние очекуваме“*, изјави Иво Бару – вработен во продавницата за органска „Добра земја“.
-Направивме споредба во цените на органски произведени зеленчук и овошје и неоргански произведени. Еден килограм органски јаболка  чинат 77 денари, килограм неоргански јаболка  чинат 32 денари, органската марула е 37 денари, додека неорганска чини 20 денари, парче органски модар патлиџан чини 35 денари, а  килограм неоргански е 59 денари.
-
-Според Националниот план за органско производство, во Македонија владее недоволна информираност на потрошувачите за бенефитите од употребата на органски земјоделски производи, недоволна информираност на трговците за постоење на субвенции за продажба на органски производи, недоволно развиен домашен пазар и непостоење на специјализирани продавници за органска храна, особено во помалите населени места. Не се доволни ниту активностите за промоција на органското производство и подигнување на свеста на потрошувачите, ниту пак количините на органски производи за континуирано снабдување на супермаркетите.
-
-Бројни истражувања говорат за позитивните ефекти од органската храна, како овошјето и зеленчукот кои содржат повеќе витамини. Органските продукти содржат помалку резидуи од тешки метали, нитрати и пестициди, издржуваат подолг рок на складирање, без да го изгубат својот квалитет и имаат влијание во заштита на природната средина. Бројни студии докажуваат дека нивото на емисија на јаглероден диоксид по хектар може да биде за 50 % пониско кај органското земјоделство во споредба со конвенционалното.
