--- content/portfolio_mk/valmira-berisha-studente-e-verber-por-ekselente.md
@@ -1,52 +1,0 @@
----
-title: Ваљмира Бериша – слепа но, извонредна студентка
-cover: '/images/uploads/valmira-berisha.jpg'
-date: 2020-02-24
-author: ['Dorina Ajdini', 'Ardi Zhuta']
-published:
- Diversity Media: 'http://diversitymedia.mk/%d0%b2%d0%b0%d1%99%d0%bc%d0%b8%d1%80%d0%b0-%d0%b1%d0%b5%d1%80%d0%b8%d1%88%d0%b0-%d1%81%d0%bb%d0%b5%d0%bf%d0%b0-%d0%bd%d0%be-%d0%b8%d0%b7%d0%b2%d0%be%d0%bd%d1%80%d0%b5%d0%b4%d0%bd%d0%b0-%d1%81%d1%82/'
----
-
-„Луѓето не гледаат со очи, туку со ум“. Оваа е мотото, кое целиот свој живот го следи Ваљмира Бериша, откога докторите и’ дијагностицирале Ретино патиа ментоса, болест од која го губи видот.
-
-Таа, нејзиниот брат и најмалата сестра се членови на шестчленото семејство од Призрен, кои боледуваат од делумно слепило. Но ова не ги спречува да го продолжат животот нормално. Ваљмира Бериша е девојка полна со енергија, која не се обесхрабрила од делумното губење на видот. Наместо тоа, таа се бори нејзината иднина да биде посветла, учејќи, за еден ден да стане некој во животот и да може да се излечи себеси, нејзината сестра и нејзиниот брат, без при тоа, да е товар на семејството. Досега изгубила 80% од видот, додека нејзиниот брат, за жал, поради недостаток на соодветни лекови, е целосно слеп.
-
-И покрај фактот што Ваљмира е прогласена за една од најдобрите ученички, завршувајќи го средното училиште во рекордно време и е вклучена во предавања за интегрирање на лица со посебни потреби, економската состојба на семејството не може да го дозволи постојаниот третман за оваа болест, како за неа, така и за нејзината помала сестра и брат. Поради тоа, нејзиниот брат целосно го изгубил видот, бидејќи лековите и третманите во странство чинат многу и ова семејство понекогаш мора да се откаже дури и од основните потреби за живеење за да го обезбеди видот на своите деца. Но, ниту овој факт не ја запира Ваљмира. Таа знае дека ќе дојде ден кога ќе работи напорно за да може да ги покрие сите трошоци.
-
-Но, како започна сето ова? Како се соочи со фактот дека нејзиниот вид не е како на другите? Ваљмира во ова интервју го објаснува грозното чувство кога разбрала дека нејзиниот живот ќе се промени. Но, како што вели, дали ќе биде на лошо или на добро, одлучува самата.
-
-Пишуваат: Дорина Ајдини и Арди Жута
-
--Кажете ни малку за вашата болест и за првите мигови кога разбравте, како се чувствувавте?
-
-„Видот на болеста е Ретино патиа ментоса, многу тешка болест. Навистина е неописливо да ви раскажувам за тогашните чувства, да ги објаснам со зборови, никој нема да  разбере. Само јас знам како ги доживеав. Во моментот сите мои соништа згаснаа. Но, за многу краток период го сменив мислењето и пристапот во ситуацијата во која се најдов. Си реков дека животот, работата и активноста нема да ги препуштам на болеста и така успеав да продолжам понатаму“.
-
--Дали тогаш се промени однесувањето на семејството или твоето друштво?
-
-„Не само што однесувањето на семејството не се смени, туку за среќа, секогаш сме се труделе да бидеме современи и во чекор со општеството и ова ни влијаеше позитивно“.
-
--Како се снаоѓате низ град, дали ве придружува некој?
-
-„Можам да кажам дека реално во градот, во 95% од случаите се снаоѓам сама. За среќа, луѓето не гледаат со очи туку со ум, човекот треба да создаде визуелна мапа во главата за да има пристап. Ова можам да го истакнам со силно убедување и задоволство дека слепите луѓе страдаат од полесна болест од луѓето со посебни потреби, бидејќи сите работи ги вршат сами, можат да пристапат до работната средина и животот, само ако имаат волја, таков е мојот пристап. Има случаи кога и мене, на пример, ме придружуваат родителите до некаде или мојата пријателка или некој друг, но, ова се ретки случаи бидејќи јас се трудам сама да се снаоѓам“.
-
--Дали имало моменти кога сте биле навредени или некој не ве разбрал?
-
-„Навреди има тука и таму. Најмногу сум била навредена во случаите кога професорите сакаа да ни го негираат правото да не‘ испрашуваат или слично. Навистина сум се чувствувала крајно скршено и уморно и тоа направи да се чувствувам лошо. Најлошите случаи беа со професорите“.
-
-Но, освен овие професори, имало и од такви, кои што полека ги вклучиле Ваљмира и нејзиниот брат во настава, станувајќи активни учесници при одговарање или разни активности. Ваљмира Бериша не се откажала од учењето, иако имала потешкотии при читање. Таа дури беше прогласена за најдобра студентка, која откога го заврши Правниот факултет, продолжи со магистерски студии по психологија.
-
--Ти си прогласена за извонредна студентка. Како го постигна тоа и кој беше начинот на учење?
-
-„Начинот на студирање беше многу интересен, бидејќи сите материјали или лекции кои требаше да ги научиме, мајка ми ги читаше, ги снимаше, а потоа сите ги ставаше на разни ЦД-а, така учев со лаптоп. Бидејќи во тоа време немавме услови, јас и брат ми бевме принудени да учиме од еден лаптоп, делејќи ги слушалките“.
-
--Дали имало неправедност или фаворизирање од професорите?
-
-„Во текот на студиите се соочувавме со тешкотии, како небрежност на професорите, нивните предрасуди дека ние сме различни од другите и дека немаме доволно способности да одговараме на испитите. Но, од друга страна, не треба да ги занемариме и професорите кои ни ги дадоа сите можности да одговараме и никогаш не ги разочаравме. што ни дадоа шанса“.
-
--Дали потребните лекови се скапи и каде ги наоѓате?
-
-„Што се однесува до лековите, ние постојано земаме лекови, посебно витамин А, витамини кои лекарот ги препишува за постојана употреба, бидејќи е не можат да се земаат преку редовната исхрана. Лековите чинат многу. Донекаде е лесно да се најдат. Но, постојано се трудиме, никогаш не сме се предале. И средствата што ги добиваме, најголемиот дел ги користиме за лечење, третмани и лекови и ова ни оневозможува луксузен па дури и нормален живот“.
-
--Дали има шанса за потполно оздравување?
-
-„Нашата болест навистина бара постојано лекување или третмани, како на пример, треба да се следи состојбата за да не се исушат крвните садови кои кај нас, на жалост, се стеснуваат, така што ние треба да сме подготвени во секој момент и да ги лечиме. За жал, многу мали ни се можностите да ги следиме овие третмани и како резултат на тоа, малиот брат целосно го изгуби видот. Јас се трудам преку други можности да ги следам третманите, но малку е неостварливо, бидејќи навистина бараат доста средства, посебно надвор од државата, главно во Турција. Има можност да се заштити видот со третмани и заодредена сума на пари.
